modularize code
Reorganized electron/main.js to use a modular structure, moving functionality into separate files.
