+++ electron/main.js
@@ -1,5 +1,5 @@
-import{app as D,BrowserWindow as Ne,ipcMain as x,dialog as ye,protocol as _e,shell as Re}from"electron";import{createRequire as Qe}from"node:module";import p from"node:path";import{fileURLToPath as Ye,pathToFileURL as Et}from"node:url";import Z from"node:https";import y from"node:fs";import Ze from"node:zlib";import{spawn as Q}from"node:child_process";import et from"node:os";import P from"node:fs";import b from"node:path";import Fe from"node:crypto";import{pipeline as je}from"node:stream";import{promisify as We}from"node:util";import ie from"node:https";var Be=We(je);function ae(n){return String(n||"").replace(/\\/g,"/").replace(/^\/+/,"")}function xe(n){return String(n||"").replace(/\\/g,"/").toLowerCase()}var K=new Map,Pe=new ie.Agent({keepAlive:!0,maxSockets:64,maxFreeSockets:32,keepAliveMsecs:6e4,scheduling:"fifo"});function O(n){return new Promise((t,r)=>{let e=Fe.createHash("sha256"),o=P.createReadStream(n);o.on("error",r),o.on("end",()=>t(e.digest("hex"))),o.on("data",a=>e.update(a))})}function Ce(n){P.mkdirSync(n,{recursive:!0})}function qe(n){return new Promise((t,r)=>{let e=ie.get(n,{agent:Pe},o=>{if(o.statusCode!==200){r(new Error(`HTTP ${o.statusCode} for ${n}`)),o.resume();return}let a="";o.setEncoding("utf8"),o.on("data",s=>a+=s),o.on("end",()=>{try{t(JSON.parse(a))}catch(s){r(s)}})});e.on("error",r),e.setTimeout(3e4,()=>{e.destroy(),r(new Error("JSON fetch timeout"))})})}function j(n){return new Promise(t=>setTimeout(t,n))}function Ae(){return{cancelled:!1,requests:new Set}}function ce(n){if(n){n.cancelled=!0;for(let t of n.requests)try{t.destroy(new Error("cancelled"))}catch{}n.requests.clear()}}var He=new Set(["ECONNRESET","ETIMEDOUT","EAI_AGAIN","ECONNABORTED","ENETRESET","EHOSTUNREACH","ENETUNREACH","EPIPE","ENOTFOUND","ECONNREFUSED","EHOSTDOWN","ENETDOWN"]);async function J(n,t,r,e=1,o,a=0,s=0){return new Promise((i,f)=>{Ce(b.dirname(t));let c=P.createWriteStream(t,{flags:a>0?"a":"w"});if(o?.cancelled)return c.close(()=>{}),f(new Error("cancelled"));let u={agent:Pe};a>0&&(u.headers={Range:`bytes=${a}-`});let d=ie.get(n,u,l=>{if(l.statusCode!==200&&!(a>0&&l.statusCode===206)){if(c.close(()=>{}),a>0&&l.statusCode===200){try{P.unlinkSync(t)}catch{}let m=Math.min(2e3*e,1e4);l.resume();try{d.destroy(new Error("restart:range-ignored"))}catch{}return j(m).then(()=>i(J(n,t,r,e+1,o,0,s)))}if(P.unlink(t,()=>{}),l.statusCode&&[429,500,502,503,504].includes(l.statusCode)&&e<5){let m=Math.min(2e3*e,1e4);return l.resume(),j(m).then(()=>i(J(n,t,r,e+1,o,a,s)))}f(new Error(`HTTP ${l.statusCode} for ${n}`)),l.resume();return}let g=Number(l.headers["content-length"]||0),v=s>0?s:a>0?a+g:g,h=0,k=Date.now(),w=e<=2?3e4:45e3,E=setInterval(()=>{if(!o?.cancelled&&Date.now()-k>w)try{let m=new Error("stalled");m.code="ETIMEDOUT",d.destroy(m)}catch{}},1e4);l.on("data",m=>{h+=m.length,k=Date.now(),r&&r(Math.min(a+h,v||a+h),v||0)}),l.on("aborted",async()=>{let m=new Promise(C=>c.close(()=>C()));try{clearInterval(E)}catch{}try{await m}catch{}if(e<5){let C=Math.min(2e3*e,1e4);return j(C).then(()=>i(J(n,t,r,e+1,o,a+h,s)))}f(new Error("response aborted"))}),Be(l,c).then(()=>{try{clearInterval(E)}catch{}i()}).catch(m=>{try{clearInterval(E)}catch{}f(m)})});if(o?.requests.add(d),d.on("socket",l=>{try{l.setKeepAlive(!0,6e4),l.setNoDelay(!0),l.setRecvBufferSize&&l.setRecvBufferSize(64*1024),l.setSendBufferSize&&l.setSendBufferSize(64*1024)}catch{}}),d.on("close",()=>o?.requests.delete(d)),d.setTimeout(9e4,()=>{try{let l=new Error("Request timeout");l.code="ETIMEDOUT",d.destroy(l)}catch{}}),o?.cancelled)try{d.destroy(new Error("cancelled"))}catch{}d.on("error",async l=>{let g=l?.code,v=e<8&&g&&He.has(String(g)),h=new Promise(k=>{try{c.close(()=>k())}catch{k()}});try{await h}catch{}if(v){let k=Math.min(1e3*Math.pow(1.5,e),15e3),w=Math.random()*1e3,E=k+w,m=a;try{m=P.statSync(t).size}catch{}return j(E).then(()=>i(J(n,t,r,e+1,o,m,s)))}f(l)})})}async function Oe(n,t,r){try{let e=P.statSync(n);return r&&Number(r)>0&&e.size!==Number(r)?!1:(await O(n)).toLowerCase()===String(t).toLowerCase()}catch{return!1}}async function be(n){let t=`${n.replace(/\/$/,"")}/checksums.json`;return qe(t)}async function Je(n,t,r,e,o=4,a){let s=ae(t.path).split("/").join(b.sep),i=b.join(r,s);if(Ce(b.dirname(i)),await Oe(i,t.checksum,t.size))return e("progress:skip",{path:t.path}),{downloadComplete:!0,verificationComplete:!0,targetPath:i};if(a?.cancelled)throw new Error("cancelled");if(Array.isArray(t.parts)&&t.parts.length>0){let c=t.parts.length,u=new Array(c),d=0,l=new Set;async function g(){for(;;){let h=d++;if(h>=c)return;if(l.has(h))continue;l.add(h);let k=t.parts[h],w=ae(k.path),E=`${n.replace(/\/$/,"")}/${w}`,m=`${i}.part${h}`;try{if(P.existsSync(m))if((await O(m)).toLowerCase()===String(k.checksum).toLowerCase()){e("progress:part",{path:t.path,part:h,totalParts:c,received:Number(k.size||0),total:Number(k.size||0)}),u[h]=m;continue}else try{P.unlinkSync(m)}catch{}}catch{}let C=0;for(;;){if(a?.cancelled)throw new Error("cancelled");C+=1;let A=0,$=0;try{let I=0;try{I=P.statSync(m).size}catch{}let N=Number(k.size||0);if(N>0&&I>N)try{P.unlinkSync(m),I=0}catch{}await J(E,m,(L,W)=>{let B=Math.max(0,L-$);$=L;try{B>0&&e("progress:bytes",{delta:B})}catch{}A+=B;let H=N||W||0;e("progress:part",{path:t.path,part:h,totalParts:c,received:Math.min(L,H||L),total:H})},1,a,I,N)}catch{try{P.unlinkSync(m)}catch{}try{A>0&&e("progress:bytes",{delta:-A})}catch{}try{e("progress:part:reset",{path:t.path,part:h,totalParts:c})}catch{}let N=Math.min(2e3*C,1e4);await j(N);continue}if((await O(m)).toLowerCase()===String(k.checksum).toLowerCase())break;try{P.unlinkSync(m)}catch{}try{A>0&&e("progress:bytes",{delta:-A})}catch{}try{e("progress:part:reset",{path:t.path,part:h,totalParts:c})}catch{}let _=Math.min(2e3*C,1e4);await j(_)}u[h]=m}}let v=Array.from({length:Math.max(1,o)},()=>g());return await Promise.all(v),{downloadComplete:!0,verificationComplete:!1,targetPath:i,mergeAndVerifyAsync:async()=>{e("progress:merge:start",{path:t.path,parts:u.length});let h=P.createWriteStream(i);for(let w=0;w<u.length;w++){let E=u[w];e("progress:merge:part",{path:t.path,part:w,totalParts:u.length}),await new Promise((m,C)=>{let A=P.createReadStream(E);A.on("error",C),h.on("error",C),A.on("end",()=>{try{P.unlinkSync(E)}catch{}m()}),A.pipe(h,{end:!1})})}await new Promise((w,E)=>{h.on("error",E),h.on("finish",w),h.end()});try{e("progress:merge:done",{path:t.path})}catch{}if(e("progress:verify",{path:t.path}),(await O(i)).toLowerCase()!==String(t.checksum).toLowerCase())throw new Error(`Checksum mismatch for ${t.path}`);try{let w=P.readdirSync(b.dirname(i)),E=b.basename(i);w.forEach(m=>{if(m.startsWith(E+".part"))try{P.unlinkSync(b.join(b.dirname(i),m))}catch{}})}catch{}return{verificationComplete:!0,targetPath:i,filePath:t.path}}}}else{let c=ae(t.path),u=`${n.replace(/\/$/,"")}/${c}`,d=0;for(;;){d+=1;let l=0,g=0,v=`${i}.download`,h=0;try{h=P.statSync(v).size}catch{}let k=Number(t.size||0);if(k>0&&h>k)try{P.unlinkSync(v),h=0}catch{}if(await J(u,v,(E,m)=>{let C=Math.max(0,E-g);g=E;try{C>0&&e("progress:bytes",{delta:C})}catch{}l+=C;let A=k||m||0;e("progress:file",{path:t.path,received:Math.min(E,A||E),total:A})},1,a,h,k),(await O(v)).toLowerCase()===String(t.checksum).toLowerCase())break;try{P.unlinkSync(v)}catch{}try{l>0&&e("progress:bytes",{delta:-l})}catch{}if(d>=2)throw new Error(`Checksum mismatch for ${t.path}`)}try{P.unlinkSync(i)}catch{}P.renameSync(`${i}.download`,i)}if(e("progress:verify",{path:t.path}),(await O(i)).toLowerCase()!==String(t.checksum).toLowerCase())throw new Error(`Checksum mismatch for ${t.path}`);return{downloadComplete:!0,verificationComplete:!0,targetPath:i}}async function De(n,t,r,e,o=!1,a=4,s=4,i,f){let c=new Set,u=[];for(let E of t.files||[]){if(!(o||!E.optional))continue;let m=xe(E.path);m&&(c.has(m)||(c.add(m),u.push(E)))}let d=u.filter(E=>!(Array.isArray(E.parts)&&E.parts.length>0)),l=u.filter(E=>Array.isArray(E.parts)&&E.parts.length>0),g=d.concat(l),v=g.length,h=0,k=g.reduce((E,m)=>{let C=Number(m.size||0);return C>0?E+C:Array.isArray(m.parts)&&m.parts.length?E+m.parts.reduce((A,$)=>A+Number($.size||0),0):E},0);try{e("progress:bytes:total",{totalBytes:k})}catch{}async function w(E,m,C,A=!1){let $=0,M=[];async function _(){for(;;){for(;f&&f();)if(await j(100),i?.cancelled)return;let N=$++;if(N>=E.length)return;let L=E[N],W=m+N,B=xe(L.path),H=async()=>Je(n,L,r,e,s,i),Se=!1,ee=K.get(B);ee?Se=!0:(e("progress:start",{index:W,total:v,path:L.path,completed:h}),ee=(async()=>{try{let F=await H();if(A&&F?.mergeAndVerifyAsync){let Ee=F.mergeAndVerifyAsync().then(R=>(R?.filePath&&(h+=1,e("progress:done",{index:W,total:v,path:R.filePath,completed:h})),R)).catch(R=>{if(String(R?.message||R||"error").includes("cancelled"))throw R;try{let U=b.join(r,L.path.replace(/\\/g,b.sep));try{P.unlinkSync(U)}catch{}try{let T=P.readdirSync(b.dirname(U)),z=b.basename(U);T.forEach(te=>{if(te.startsWith(z+".part"))try{P.unlinkSync(b.join(b.dirname(U),te))}catch{}})}catch{}return H().then(T=>T.mergeAndVerifyAsync?T.mergeAndVerifyAsync().then(z=>(z?.filePath&&(h+=1,e("progress:done",{index:W,total:v,path:z.filePath,completed:h})),z)):T)}catch(U){try{e("progress:error",{path:L.path,message:String(U?.message||U)})}catch{}throw U}});return M.push(Ee),{...F,skipProgressDone:!0}}return F}catch(F){if(String(F?.message||F||"error").includes("cancelled"))throw F;try{let R=b.join(r,L.path.replace(/\\/g,b.sep));try{P.unlinkSync(R)}catch{}let G=await H();if(A&&G?.mergeAndVerifyAsync){let U=G.mergeAndVerifyAsync().then(T=>(T?.filePath&&(h+=1,e("progress:done",{index:W,total:v,path:T.filePath,completed:h})),T)).catch(T=>{try{let z=b.join(r,L.path.replace(/\\/g,b.sep)),te=P.readdirSync(b.dirname(z)),ze=b.basename(z);te.forEach(ve=>{if(ve.startsWith(ze+".part"))try{P.unlinkSync(b.join(b.dirname(z),ve))}catch{}})}catch{}try{e("progress:error",{path:L.path,message:String(T?.message||T)})}catch{}throw T});return M.push(U),{...G,skipProgressDone:!0}}return G}catch(R){try{e("progress:error",{path:L.path,message:String(R?.message||R)})}catch{}throw R}}})(),K.set(B,ee));let ke;try{ke=await ee}finally{Se||K.delete(B)}ke?.skipProgressDone||(h+=1,e("progress:done",{index:W,total:v,path:L.path,completed:h})),await j(10)}}let I=Array.from({length:Math.max(1,C)},()=>_());await Promise.all(I),M.length>0&&await Promise.all(M)}if(await w(d,0,Math.max(1,a)),K.size>0)try{await Promise.all([...K.values()])}catch{}await w(l,d.length,1,!0)}import le from"node:fs";import Ge from"node:path";import{app as Ke}from"electron";var $e=Ke.getPath("userData"),Le=Ge.join($e,"settings.json");function de(){try{let n=le.readFileSync(Le,"utf-8");return JSON.parse(n)}catch{return{}}}function Xe(n){le.mkdirSync($e,{recursive:!0}),le.writeFileSync(Le,JSON.stringify(n,null,2),"utf-8")}function he(n,t=void 0){return de()[n]??t}function re(n,t){let r=de();r[n]=t,Xe(r)}function ue(){return de()}var Ue=Qe(import.meta.url),{autoUpdater:V}=Ue("electron-updater"),Te=Ue("electron-log"),tt=Ye(import.meta.url),X=p.dirname(tt);var S,q=null,fe=new Set,ne=new Map,oe=new Map,pe="r5v",se=[];function Ve(n){try{return(n||[]).filter(t=>typeof t=="string"&&t.startsWith(`${pe}://`))}catch{return[]}}function ge(n){try{let t=new URL(n);if(t.hostname==="mod"&&t.pathname==="/install"){let r=t.searchParams.get("name")||"",e=t.searchParams.get("version")||"",o=t.searchParams.get("downloadUrl")?.split(",")||"";if(S&&S.webContents&&!S.webContents.isLoading())try{S.webContents.send("deeplink:mod-install",{url:n,name:r,version:e,downloadUrls:o})}catch{}else se.push(n)}}catch{}}function rt(){for(;se.length;){let n=se.shift();ge(n)}}var nt=D.requestSingleInstanceLock();nt?D.on("second-instance",(n,t)=>{try{Ve(t).forEach(e=>ge(e)),S&&(S.isMinimized()&&S.restore(),S.focus())}catch{}}):D.quit();D.on("open-url",(n,t)=>{n.preventDefault(),ge(t)});async function Me(){let n=he("windowState",{width:1150,height:800,x:void 0,y:void 0,isMaximized:!1});S=new Ne({width:n.width,height:n.height,x:n.x,y:n.y,minWidth:1150,minHeight:800,frame:!1,titleBarStyle:"hidden",webPreferences:{preload:p.join(X,"preload.cjs"),contextIsolation:!0,nodeIntegration:!1},show:!1,icon:p.join(X,"..","public","icon.png")});let t=process.env.VITE_DEV_SERVER_URL;if(t)await S.loadURL(t);else{let s=p.join(X,"..","dist","index.html");await S.loadFile(s)}S.webContents.setWindowOpenHandler(({url:s})=>{try{Re.openExternal(s)}catch{}return{action:"deny"}}),S.webContents.on("will-navigate",(s,i)=>{let f=i.startsWith("file:"),c=!!process.env.VITE_DEV_SERVER_URL&&i.startsWith(String(process.env.VITE_DEV_SERVER_URL));if(!f&&!c){s.preventDefault();try{Re.openExternal(i)}catch{}}});let r=()=>{S&&(S.isVisible()||S.show(),S.isFocused()||S.focus())};n.isMaximized&&S.maximize(),S.once("ready-to-show",r),S.webContents.once("did-finish-load",r);let e=setTimeout(r,1e3);S.on("show",()=>clearTimeout(e));function o(){if(!S)return;let s=S.getBounds(),i=S.isMaximized();re("windowState",{width:s.width,height:s.height,x:s.x,y:s.y,isMaximized:i})}S.on("resize",o),S.on("move",o),S.on("maximize",o),S.on("unmaximize",o),(!!process.env.VITE_DEV_SERVER_URL||process.env.NODE_ENV==="development")&&S.webContents.openDevTools({mode:"detach"}),S.webContents.on("did-fail-load",(s,i,f,c)=>{console.error("Load failed",{errorCode:i,errorDesc:f,validatedURL:c}),ye.showErrorBox("Load failed",`${f} (code ${i})
-URL: ${c}`),S.show()})}D.on("window-all-closed",()=>{process.platform!=="darwin"&&D.quit()});D.whenReady().then(()=>{try{process.defaultApp&&process.platform==="win32"?D.setAsDefaultProtocolClient(pe,process.execPath,[process.argv[1]]):D.setAsDefaultProtocolClient(pe)}catch{}let n=p.resolve(D.getAppPath(),".."),t=p.join(n,"cache");_e.registerFileProtocol("r5v",(e,o)=>{try{let a=new URL(e.url),s=decodeURIComponent(a.pathname).replace(/^\//,""),i=p.normalize(p.join(t,s));o({path:i})}catch{o({error:-6})}});let r=p.join(X,"..","dist");_e.interceptFileProtocol("file",(e,o)=>{try{let a=new URL(e.url),s=decodeURIComponent(a.pathname),i=s.replace(/\\/g,"/"),f=i.indexOf("/_astro/");if(f!==-1){let c=i.substring(f+8),u=p.join(r,"_astro",c);return o({path:u})}if(i.endsWith("/favicon.svg")||/\/favicon\.svg$/i.test(i)){let c=p.join(r,"favicon.svg");return o({path:c})}if(i.startsWith("/")){let c=i.replace(/^\/+/,""),u=p.join(r,c);try{return y.accessSync(u),o({path:u})}catch{}}return o({path:s})}catch{return o({error:-6})}}),Me();try{Ve(process.argv).forEach(e=>se.push(e))}catch{}try{S?.webContents?.once("did-finish-load",rt)}catch{}try{V.autoDownload=!1;try{V.logger=Te,Te.transports.file.level="info"}catch{}V.on("error",e=>{try{S?.webContents.send("update:error",{message:String(e?.stack||e?.message||e)})}catch{}}),V.on("update-available",e=>{try{S?.webContents.send("update:available",e)}catch{}}),V.on("update-not-available",e=>{try{S?.webContents.send("update:not-available",e)}catch{}}),V.on("download-progress",e=>{try{S?.webContents.send("update:download-progress",e)}catch{}}),V.on("update-downloaded",e=>{try{S?.webContents.send("update:downloaded",e)}catch{}})}catch{}D.on("activate",()=>{Ne.getAllWindows().length===0&&Me()})});x.handle("update:check",async()=>{try{return{ok:!0,result:(await V.checkForUpdates())?.updateInfo||null}}catch(n){return{ok:!1,error:String(n?.message||n)}}});x.handle("update:download",async()=>{try{return await V.downloadUpdate(),{ok:!0}}catch(n){return{ok:!1,error:String(n?.message||n)}}});x.handle("update:quitAndInstall",async()=>{try{return setImmediate(()=>V.quitAndInstall(!1,!0)),{ok:!0}}catch(n){return{ok:!1,error:String(n?.message||n)}}});x.handle("app:getVersion",async()=>{try{return D.getVersion()}catch{return"0.0.0"}});x.handle("app:getBaseDir",async()=>{try{return p.resolve(D.getAppPath(),"..")}catch{return""}});x.handle("app:getLauncherInstallRoot",async()=>{try{let n=process.env.LOCALAPPDATA||p.join(D.getPath("home"),"AppData","Local");return p.join(n,"Programs","r5vlauncher")}catch{return""}});function ot(n){try{let t=p.join(n,"mods.vdf"),r=y.readFileSync(t,"utf-8"),e={},o=/"([^"]+)"\s*"([01])"/g,a;for(;a=o.exec(r);){let s=a[1];s&&s!=="ModList"&&(e[s]=a[2]==="1")}return e}catch{return{}}}function we(n){try{let t=p.join(n,"mods.vdf"),r=y.readFileSync(t,"utf-8"),e={},o=[],a=/"([^"]+)"\s*"([01])"/g,s;for(;s=a.exec(r);){let i=s[1];!i||i==="ModList"||(i in e||o.push(i),e[i]=s[2]==="1")}return{map:e,order:o}}catch{return{map:{},order:[]}}}function me(n){try{let t=y.readFileSync(n,"utf-8"),r=t.match(/"id"\s*"([^"]+)"/i),e=t.match(/"name"\s*"([^"]+)"/i),o=r?r[1]:null,a=e?e[1]:null;return{id:o,name:a}}catch{return{id:null,name:null}}}function st(n,t){let r=['"ModList"',"{"];for(let[e,o]of Object.entries(t))r.push(`	"${e}"		"${o?"1":"0"}"`);r.push("}"),y.mkdirSync(n,{recursive:!0}),y.writeFileSync(p.join(n,"mods.vdf"),r.join(`
-`),"utf-8")}function Ie(n,t,r){let e=new Set,o=['"ModList"',"{"],a=s=>{o.push(`	"${s}"		"${r[s]?"1":"0"}"`),e.add(s)};for(let s of Array.isArray(t)?t:[])Object.prototype.hasOwnProperty.call(r,s)&&!e.has(s)&&a(s);for(let s of Object.keys(r))e.has(s)||a(s);o.push("}"),y.mkdirSync(n,{recursive:!0}),y.writeFileSync(p.join(n,"mods.vdf"),o.join(`
-`),"utf-8")}x.handle("mods:listInstalled",async(n,{installDir:t})=>{try{let r=p.join(t,"mods"),{map:e,order:o}=we(r),a=[],s=y.existsSync(r)?y.readdirSync(r,{withFileTypes:!0}):[];for(let f of s){if(!f.isDirectory())continue;let c=p.join(r,f.name),u=p.join(c,"manifest.json"),d=p.join(c,"mod.vdf"),l=p.join(c,"icon.png"),g=null;try{g=JSON.parse(y.readFileSync(u,"utf-8"))}catch{}let v=me(d),h=v.id||g?.name||f.name,k=null;try{k=`data:image/png;base64,${y.readFileSync(l).toString("base64")}`}catch{}a.push({id:h,name:g?.name||v.name||f.name,folder:f.name,version:g?.version_number||null,description:g?.description||"",enabled:h?!!e[h]:!1,hasManifest:y.existsSync(u),iconDataUrl:k})}let i=new Map(o.map((f,c)=>[f,c]));return a.sort((f,c)=>{let u=f.id!=null&&i.has(f.id)?i.get(f.id):Number.MAX_SAFE_INTEGER,d=c.id!=null&&i.has(c.id)?i.get(c.id):Number.MAX_SAFE_INTEGER;return u!==d?u-d:String(f.name||"").localeCompare(String(c.name||""))}),{ok:!0,mods:a}}catch(r){return{ok:!1,error:String(r?.message||r)}}});x.handle("mods:setEnabled",async(n,{installDir:t,name:r,enabled:e})=>{try{let o=p.join(t,"mods"),{map:a}=we(o);return a[r]=!!e,Ie(o,void 0,a),{ok:!0}}catch(o){return{ok:!1,error:String(o?.message||o)}}});x.handle("mods:reorder",async(n,{installDir:t,orderIds:r})=>{try{let e=p.join(t,"mods"),{map:o,order:a}=we(e),s=Array.isArray(r)?r.filter(u=>typeof u=="string"&&u&&Object.prototype.hasOwnProperty.call(o,u)):[],i=new Set(s),f=a.filter(u=>Object.prototype.hasOwnProperty.call(o,u)&&!i.has(u)),c=[...s,...f];return Ie(e,c,o),{ok:!0}}catch(e){return{ok:!1,error:String(e?.message||e)}}});x.handle("mods:uninstall",async(n,{installDir:t,folder:r})=>{try{let e=p.join(t,"mods"),o=p.join(e,r);return y.rmSync(o,{recursive:!0,force:!0}),{ok:!0}}catch(e){return{ok:!1,error:String(e?.message||e)}}});x.handle("mods:fetchAll",async(n,{query:t})=>{let r="https://thunderstore.io/c/r5valkyrie/api/v1/package-listing-index/",e=a=>new Promise((s,i)=>{Z.get(a,{headers:{"User-Agent":"R5Valkyrie-Launcher/1.0",Accept:"*/*"}},c=>{if(c.statusCode&&c.statusCode>=300&&c.statusCode<400&&c.headers.location){c.resume();let d=c.headers.location.startsWith("http")?c.headers.location:new URL(c.headers.location,a).toString();return s(e(d))}if(c.statusCode!==200)return c.resume(),i(new Error(`HTTP ${c.statusCode}`));let u=[];c.on("data",d=>u.push(Buffer.isBuffer(d)?d:Buffer.from(d))),c.on("end",()=>s(Buffer.concat(u)))}).on("error",i)}),o=a=>{try{return Ze.gunzipSync(a)}catch{return a}};try{let a=await e(r),s=JSON.parse(o(a).toString("utf8")),i=Array.isArray(s)?s:[],f=6,c=[],u=0;await Promise.all(Array.from({length:f}).map(async()=>{for(;u<i.length;){let l=u++,g=i[l];try{let v=await e(g),h=o(v).toString("utf8"),k=JSON.parse(h);Array.isArray(k)&&c.push(...k)}catch{}}}));let d=c;if(t&&String(t).trim()){let l=String(t).toLowerCase();d=d.filter(g=>String(g?.name||"").toLowerCase().includes(l)||String(g?.full_name||"").toLowerCase().includes(l))}return{ok:!0,mods:d}}catch(a){try{let i=await e("https://thunderstore.io/c/r5valkyrie/api/v1/package/"),f=JSON.parse(i.toString("utf8")),c=Array.isArray(f)?f:[];if(t&&String(t).trim()){let u=String(t).toLowerCase();c=c.filter(d=>String(d?.name||"").toLowerCase().includes(u)||String(d?.full_name||"").toLowerCase().includes(u))}return{ok:!0,mods:c}}catch(s){return{ok:!1,error:String(s?.message||a?.message||s||a)}}}});x.handle("mods:install",async(n,{installDir:t,name:r,downloadUrl:e})=>{if(!e.startsWith("https://thunderstore.io"))return{ok:!1,error:"Download URL does not point to Thunderstore"};try{let o=String(r||"").trim();if(!o)return{ok:!1,error:"Invalid mod name"};if(fe.has(o))return{ok:!0};fe.add(o);let a=p.join(t,"mods");y.mkdirSync(a,{recursive:!0});let s=p.join(a,`.__mod_${Date.now()}.zip`),i=(d,l=0)=>new Promise((g,v)=>{if(l>5)return v(new Error("Too many redirects"));let h=y.createWriteStream(s);Z.get(d,{headers:{"User-Agent":"R5Valkyrie-Launcher/1.0",Accept:"application/octet-stream,*/*;q=0.8",Referer:"https://thunderstore.io/",Connection:"keep-alive"}},w=>{if(w.statusCode&&w.statusCode>=300&&w.statusCode<400&&w.headers.location){w.resume(),h.close(()=>{});try{y.unlinkSync(s)}catch{}let $=w.headers.location.startsWith("http")?w.headers.location:new URL(w.headers.location,d).toString();return g(i($,l+1))}if(w.statusCode!==200){h.close(()=>{});try{y.unlinkSync(s)}catch{}return w.resume(),v(new Error(`HTTP ${w.statusCode}`))}let E=Number(w.headers["content-length"]||0),m=0,C=Date.now(),A=$=>{try{n?.sender?.send("mods:progress",{key:r,phase:$,received:m,total:E})}catch{}};w.on("data",$=>{m+=Buffer.isBuffer($)?$.length:Buffer.byteLength(String($));let M=Date.now();M-C>150&&(C=M,A("downloading"))}),w.pipe(h),h.on("finish",()=>{A("downloading"),h.close(g)})}).on("error",w=>{try{y.unlinkSync(s)}catch{}v(w)})});await i(e);try{n?.sender?.send("mods:progress",{key:r,phase:"extracting"})}catch{}let f=p.join(a,r);try{y.rmSync(f,{recursive:!0,force:!0})}catch{}y.mkdirSync(f,{recursive:!0}),await new Promise((d,l)=>{let g=Q("powershell.exe",["-NoProfile","-NonInteractive","-Command",`Expand-Archive -LiteralPath "${s}" -DestinationPath "${f}" -Force`],{stdio:"ignore"});g.on("exit",v=>v===0?d():l(new Error(`Expand-Archive failed ${v}`))),g.on("error",l)});try{y.unlinkSync(s)}catch{}let c=null;try{let d=p.join(f,"mod.vdf");if(y.existsSync(d))c=me(d).id;else{let l=y.readdirSync(f,{withFileTypes:!0});for(let g of l)if(g.isDirectory()){let v=p.join(f,g.name,"mod.vdf");if(y.existsSync(v)&&(c=me(v).id,c))break}}}catch{}let u=ot(a);u[c||r]=!0,st(a,u);try{n?.sender?.send("mods:progress",{key:r,phase:"done"})}catch{}return{ok:!0}}catch(o){return{ok:!1,error:String(o?.message||o)}}finally{try{fe.delete(String(r||""))}catch{}}});x.handle("mods:watch",async(n,{installDir:t})=>{try{let r=p.join(t,"mods");if(!y.existsSync(r))return{ok:!0};let e=r;if(ne.has(e))return{ok:!0};let o=y.watch(r,{persistent:!0},(a,s)=>{let i=e;clearTimeout(oe.get(i));let f=setTimeout(()=>{try{S?.webContents?.send("mods:changed",{installDir:t})}catch{}},300);oe.set(i,f)});return ne.set(e,o),{ok:!0}}catch(r){return{ok:!1,error:String(r?.message||r)}}});x.handle("mods:unwatch",async(n,{installDir:t})=>{try{let e=p.join(t,"mods"),o=ne.get(e);if(o){try{o.close()}catch{}ne.delete(e)}let a=oe.get(e);return a&&(clearTimeout(a),oe.delete(e)),{ok:!0}}catch(r){return{ok:!1,error:String(r?.message||r)}}});x.handle("mods:iconDataUrl",async(n,{installDir:t,folder:r})=>{try{let e=p.join(t,"mods",r,"icon.png");return await y.promises.access(e),{ok:!0,dataUrl:`data:image/png;base64,${(await y.promises.readFile(e)).toString("base64")}`}}catch(e){return{ok:!1,error:String(e?.message||e)}}});x.handle("select-directory",async()=>{let n=await ye.showOpenDialog({properties:["openDirectory","createDirectory"]});return n.canceled||n.filePaths.length===0?null:n.filePaths[0]});x.handle("settings:get",()=>ue());x.handle("settings:set",(n,{key:t,value:r})=>re(t,r));x.handle("download:checksums",async(n,{baseUrl:t})=>be(t));x.handle("download:all",async(n,{baseUrl:t,checksums:r,installDir:e,includeOptional:o,concurrency:a,partConcurrency:s,channelName:i,mode:f})=>{let c=(u,d)=>n.sender.send(u,d);try{if(!(String(f||"install").toLowerCase()==="install")){let d=p.join(e,"mods","mods.vdf");if(y.existsSync(d)){let g=(Array.isArray(r?.files)?r.files:[]).filter(v=>{try{return String(v?.path||v?.name||"").replace(/\\/g,"/").toLowerCase()!=="mods/mods.vdf"}catch{return!0}});r={...r||{},files:g}}}}catch{}q=Ae(),Y=!1;try{await De(t,r,e,c,!!o,Number(a)||4,Number(s)||4,q,()=>Y);try{let u=he("channels",{})||{};u[String(i||"default")]={installDir:e,gameVersion:r?.game_version||null,gameBaseUrl:t,lastUpdatedAt:Date.now()},re("channels",u)}catch(u){console.error("Failed to persist channel info",u)}}finally{q=null}return!0});x.handle("default-install-dir",(n,{channelName:t})=>{let r="";if(process.platform==="win32"){let e=process.env.LOCALAPPDATA||p.join(D.getPath("home"),"AppData","Local");r=p.join(e,"Programs","R5VLibrary")}else if(process.platform==="linux"){let e=p.join(D.getPath("home"),"Games");r=p.join(e,"R5VLibrary")}return t?p.join(r,t):r});x.handle("launcher:config",async(n,{url:t})=>(e=>new Promise((o,a)=>{Z.get(e,s=>{if(s.statusCode!==200){a(new Error(`HTTP ${s.statusCode} for ${e}`)),s.resume();return}let i="";s.setEncoding("utf8"),s.on("data",f=>i+=f),s.on("end",()=>{try{o(JSON.parse(i))}catch(f){a(f)}})}).on("error",a)}))(t));x.handle("eula:get",async()=>{let n="https://playvalkyrie.org/api/eula",t=r=>new Promise((e,o)=>{Z.get(r,a=>{if(a.statusCode!==200){o(new Error(`HTTP ${a.statusCode}`)),a.resume();return}let s=[];a.on("data",i=>s.push(Buffer.isBuffer(i)?i:Buffer.from(i))),a.on("end",()=>{try{let i=Buffer.concat(s).toString("utf8");e(JSON.parse(i))}catch(i){o(i)}})}).on("error",o)});try{return{ok:!0,json:await t(n)}}catch(r){return{ok:!1,error:String(r?.message||r)}}});x.handle("video:cache",async(n,{filename:t})=>{try{let e=`https://blaze.playvalkyrie.org/video_backgrounds/${t}`,o=p.join(p.resolve(D.getAppPath(),".."),"cache"),a=p.join(o,"videos"),s=p.join(a,t),i=p.join(X,"..","dist"),f=p.join(i,t);await y.promises.mkdir(a,{recursive:!0});try{if((await y.promises.stat(s)).size>0){try{await y.promises.access(f)}catch{await y.promises.copyFile(s,f)}return`./${t}`}else await y.promises.unlink(s)}catch{}return await new Promise((c,u)=>{let d=y.createWriteStream(s),l=Z.get(e,g=>{if(g.statusCode!==200){d.close(()=>{}),y.unlink(s,()=>{}),u(new Error(`HTTP ${g.statusCode} for video ${t}`)),g.resume();return}g.pipe(d),d.on("finish",()=>d.close(c)),d.on("error",v=>{y.unlink(s,()=>u(v))})});l.on("error",g=>{d.close(()=>{}),y.unlink(s,()=>{}),u(g)}),l.setTimeout(3e4,()=>{l.destroy(),d.close(()=>{}),y.unlink(s,()=>{}),u(new Error(`Video download timeout for ${t}`))})}),await y.promises.copyFile(s,f),`./${t}`}catch(r){throw console.error(`Video cache error for ${t}:`,r),r}});x.handle("fs:is-installed-in-dir",async(n,{path:t})=>{let r=!1,e=!1;try{await y.promises.access(p.join(t,"r5apex.exe")),r=!0}catch{}try{await y.promises.access(p.join(t,"r5apex_ds.exe")),e=!0}catch{}return r||e});var Y=!1;x.handle("download:pause",async()=>{Y=!0;try{S?.webContents.send("progress:paused",{})}catch{}return!0});x.handle("download:resume",async()=>{Y=!1;try{S?.webContents.send("progress:resumed",{})}catch{}return!0});x.handle("download:cancel",async()=>{try{ce(q)}catch{}q=null,Y=!1;try{S?.webContents.send("progress:cancelled",{})}catch{}return!0});x.handle("select-file",async(n,{filters:t})=>{let r=await ye.showOpenDialog({properties:["openFile"],filters:Array.isArray(t)?t:void 0});return r.canceled||r.filePaths.length===0?null:r.filePaths[0]});x.handle("game:launch",async(n,{channelName:t,installDir:r,mode:e,argsString:o,winePrefix:a})=>{try{if(!r)throw new Error("Missing installDir");let s=String(e).toUpperCase()==="SERVER"?"r5apex_ds.exe":"r5apex.exe",i=p.join(r,s);await y.promises.access(i).catch(()=>{throw new Error(`Executable not found: ${i}`)});let c=(d=>!d||typeof d!="string"?[]:(d.match(/(?:[^\s\"]+|\"[^\"]*\")+/g)||[]).map(g=>g.replace(/^\"|\"$/g,"")))(o),u;if(process.platform==="win32")u=Q(i,c,{cwd:r,detached:!0,stdio:"ignore",env:{...process.env}});else if(process.platform==="linux"){let d=a||p.join(D.getPath("home"),"Games","R5VLibrary","wineprefix");c.unshift(i);let l=p.join(D.getPath("home"),".local","share","Steam"),g=p.join(d,"drive_c","Program Files (x86)","Steam");y.mkdirSync(g,{recursive:!0}),["GameOverlayRenderer64.dll","steamclient64.dll","steamclient.dll"].forEach(h=>{let k=p.join(l,h),w=p.join(g,h);if(!y.existsSync(k))throw new Error(`Steam is not installed correctly!
-Source file not found: ${k}`);y.existsSync(w)?console.log(`Skipped (exists): ${h}`):(y.copyFileSync(k,w),console.log(`Copied: ${h}`))}),u=Q("umu-run",c,{cwd:r,detached:!0,stdio:"ignore",env:{...process.env,WINEPREFIX:d}})}return u.unref(),{ok:!0}}catch(s){return{ok:!1,error:String(s?.message||s)}}});x.handle("fix-folder-permissions",async(n,{selectedChannel:t})=>{let r=[],e=ue().channels[t].installDir;try{if(!e)return{ok:!1,error:"No folder path provided"};let o=p.resolve(e).replace(/\//g,"\\");if(!o||o.length<3)return{ok:!1,error:`Invalid path format: ${e}`};let a=o.length>=3&&o[1]===":"?o.slice(3):o;if(/[<>:"|?*]/.test(a))return{ok:!1,error:`Path contains invalid characters: ${o}`};let f=et.userInfo().username,c=p.join(process.resourcesPath,"elevate.exe"),u=y.existsSync(c);try{y.mkdirSync(o,{recursive:!0})}catch{}let d=[{exe:"icacls",args:[o,"/grant",`${f}:F`,"/t","/q"],desc:`Grant permissions to current user (${f})`},{exe:"icacls",args:[o,"/grant","Everyone:F","/t","/q"],desc:"Grant Everyone full permissions"},{exe:"icacls",args:[o,"/grant","Users:F","/t","/q"],desc:"Grant Users full permissions"}];for(let l=0;l<d.length;l++){let{exe:g,args:v,desc:h}=d[l];try{let k=await new Promise((w,E)=>{let m,C,A;u?(C=[c,["-wait",g,...v]],A={stdio:["ignore","pipe","pipe"],windowsHide:!0}):(C=[g,v],A={stdio:["ignore","pipe","pipe"],windowsHide:!0});try{m=Q(C[0],C[1],A)}catch(_){console.error("[Permission Fix] Spawn failed:",_),E(_);return}let $="",M="";m.stdout&&m.stdout.on("data",_=>{$+=_.toString()}),m.stderr&&m.stderr.on("data",_=>{M+=_.toString()}),m.on("exit",(_,I)=>{w({code:_,stdout:$,stderr:M,signal:I})}),m.on("error",_=>{E(_)}),setTimeout(()=>{try{m.kill()}catch{}E(new Error("Command timeout"))},3e4)});if(k.code!==0){let w=k.stderr||k.stdout||"Unknown error";if(!(w.includes("duplicate")||w.includes("already")||k.code===1332)){let m=`${h} failed (code ${k.code}): ${w}`;r.push(m)}}}catch(k){let w=`${h} error: ${k.message}`;r.push(w)}}try{let l=p.join(o,"test_write.tmp");y.writeFileSync(l,"test"),y.unlinkSync(l)}catch(l){let g=`Write test failed: ${l.message}`;r.push(g);try{if(y.mkdirSync(o,{recursive:!0}),(await new Promise(h=>{let k=Q("cmd",["/c",`icacls "${o}" /grant "%USERNAME%":F /q`],{stdio:"ignore",windowsHide:!0});k.on("exit",w=>h({code:w})),k.on("error",()=>h({code:1})),setTimeout(()=>{try{k.kill()}catch{}h({code:1})},5e3)})).code===0){let h=p.join(o,"test_write2.tmp");return y.writeFileSync(h,"test"),y.unlinkSync(h),{ok:!0,warnings:r}}}catch{}return{ok:!1,error:g,details:r}}return r.length>0?{ok:!0,warnings:r}:{ok:!0}}catch(o){let a=`Unexpected error: ${o?.message||o}`;return r.push(a),{ok:!1,error:a,details:r}}});x.on("window:minimize",()=>{S?.minimize()});x.on("window:maximize",()=>{S&&(S.isMaximized()?S.unmaximize():S.maximize())});x.on("window:close",()=>{S?.close()});D.on("before-quit",()=>{if(q)try{ce(q)}catch{}});
+var fe=(r=>typeof require<"u"?require:typeof Proxy<"u"?new Proxy(r,{get:(t,n)=>(typeof require<"u"?require:t)[n]}):r)(function(r){if(typeof require<"u")return require.apply(this,arguments);throw Error('Dynamic require of "'+r+'" is not supported')});import{app as st,BrowserWindow as nr,ipcMain as O}from"electron";import{BrowserWindow as we,shell as Ut}from"electron";import dt from"node:path";import{fileURLToPath as Se,pathToFileURL as pr}from"node:url";import yt from"node:fs";import me from"node:path";import{app as ye}from"electron";var Rt=ye.getPath("userData"),Tt=me.join(Rt,"settings.json");function gt(){try{let r=yt.readFileSync(Tt,"utf-8");return JSON.parse(r)}catch{return{}}}function ge(r){yt.mkdirSync(Rt,{recursive:!0}),yt.writeFileSync(Tt,JSON.stringify(r,null,2),"utf-8")}function ct(r,t=void 0){return gt()[r]??t}function G(r,t){let n=gt();n[r]=t,ge(n)}function lt(){return gt()}var ke=Se(import.meta.url),wt=dt.dirname(ke);async function St(){let r=ct("windowState",{width:1150,height:800,x:void 0,y:void 0,isMaximized:!1}),t=new we({width:r.width,height:r.height,x:r.x,y:r.y,minWidth:1150,minHeight:800,frame:!1,titleBarStyle:"hidden",webPreferences:{preload:dt.join(wt,"preload.cjs"),contextIsolation:!0,nodeIntegration:!1},show:!1,icon:dt.join(wt,"..","public","icon.png")}),n=process.env.VITE_DEV_SERVER_URL;if(n)await t.loadURL(n);else{let a=dt.join(wt,"..","dist","index.html");await t.loadFile(a)}t.webContents.setWindowOpenHandler(({url:a})=>{try{Ut.openExternal(a)}catch{}return{action:"deny"}}),t.webContents.on("will-navigate",(a,d)=>{let c=d.startsWith("file:"),f=!!process.env.VITE_DEV_SERVER_URL&&d.startsWith(String(process.env.VITE_DEV_SERVER_URL));if(!c&&!f){a.preventDefault();try{Ut.openExternal(d)}catch{}}});let e=()=>{t&&(t.isVisible()||t.show(),t.isFocused()||t.focus())};r.isMaximized&&t.maximize(),t.once("ready-to-show",e),t.webContents.once("did-finish-load",e);let o=setTimeout(e,1e3);t.on("show",()=>clearTimeout(o));function s(){if(!t)return;let a=t.getBounds(),d=t.isMaximized();G("windowState",{width:a.width,height:a.height,x:a.x,y:a.y,isMaximized:d})}return t.on("resize",s),t.on("move",s),t.on("maximize",s),t.on("unmaximize",s),(!!process.env.VITE_DEV_SERVER_URL||process.env.NODE_ENV==="development")&&t.webContents.openDevTools({mode:"detach"}),t.webContents.on("did-fail-load",(a,d,c,f)=>{console.error("Load failed",{errorCode:d,errorDesc:c,validatedURL:f});let{dialog:u}=fe("electron");u.showErrorBox("Load failed",`${c} (code ${d})
+URL: ${f}`),t.show()}),t}function Nt(r,t){r.on("window:minimize",()=>{t()?.minimize()}),r.on("window:maximize",()=>{let n=t();n&&(n.isMaximized()?n.unmaximize():n.maximize())}),r.on("window:close",()=>{t()?.close()})}import{app as K}from"electron";var kt="r5v",ht=[];function Vt(r){try{return(r||[]).filter(t=>typeof t=="string"&&t.startsWith(`${kt}://`))}catch{return[]}}function xt(r,t){try{let n=new URL(r);if(n.hostname==="mod"&&n.pathname==="/install"){let e=n.searchParams.get("name")||"",o=n.searchParams.get("version")||"",s=n.searchParams.get("downloadUrl")?.split(",")||"",i=t();if(i&&i.webContents&&!i.webContents.isLoading())try{i.webContents.send("deeplink:mod-install",{url:r,name:e,version:o,downloadUrls:s})}catch{}else ht.push(r)}}catch{}}function Ft(r){for(;ht.length;){let t=ht.shift();xt(t,r)}}function zt(r){return K.requestSingleInstanceLock()?(K.on("second-instance",(n,e)=>{try{Vt(e).forEach(i=>xt(i,r));let s=r();s&&(s.isMinimized()&&s.restore(),s.focus())}catch{}}),K.on("open-url",(n,e)=>{n.preventDefault(),xt(e,r)}),!0):(K.quit(),!1)}function jt(){try{process.defaultApp&&process.platform==="win32"?K.setAsDefaultProtocolClient(kt,process.execPath,[process.argv[1]]):K.setAsDefaultProtocolClient(kt)}catch{}}function Bt(){try{Vt(process.argv).forEach(r=>ht.push(r))}catch{}}import{createRequire as xe}from"node:module";var qt=xe(import.meta.url),{autoUpdater:N}=qt("electron-updater"),Wt=qt("electron-log");function Ht(r){try{N.autoDownload=!1;try{N.logger=Wt,Wt.transports.file.level="info"}catch{}N.on("error",t=>{try{r()?.webContents.send("update:error",{message:String(t?.stack||t?.message||t)})}catch{}}),N.on("update-available",t=>{try{r()?.webContents.send("update:available",t)}catch{}}),N.on("update-not-available",t=>{try{r()?.webContents.send("update:not-available",t)}catch{}}),N.on("download-progress",t=>{try{r()?.webContents.send("update:download-progress",t)}catch{}}),N.on("update-downloaded",t=>{try{r()?.webContents.send("update:downloaded",t)}catch{}})}catch{}}function Mt(r){r.handle("update:check",async()=>{try{return{ok:!0,result:(await N.checkForUpdates())?.updateInfo||null}}catch(t){return{ok:!1,error:String(t?.message||t)}}}),r.handle("update:download",async()=>{try{return await N.downloadUpdate(),{ok:!0}}catch(t){return{ok:!1,error:String(t?.message||t)}}}),r.handle("update:quitAndInstall",async()=>{try{return setImmediate(()=>N.quitAndInstall(!1,!0)),{ok:!0}}catch(t){return{ok:!1,error:String(t?.message||t)}}})}import Ot from"node:https";import P from"node:fs";import C from"node:path";import Pe from"node:zlib";import{spawn as Ee}from"node:child_process";var Pt=new Set,ut=new Map,pt=new Map;function ve(r){try{let t=C.join(r,"mods.vdf"),n=P.readFileSync(t,"utf-8"),e={},o=/"([^"]+)"\s*"([01])"/g,s;for(;s=o.exec(n);){let i=s[1];i&&i!=="ModList"&&(e[i]=s[2]==="1")}return e}catch{return{}}}function vt(r){try{let t=C.join(r,"mods.vdf"),n=P.readFileSync(t,"utf-8"),e={},o=[],s=/"([^"]+)"\s*"([01])"/g,i;for(;i=s.exec(n);){let a=i[1];!a||a==="ModList"||(a in e||o.push(a),e[a]=i[2]==="1")}return{map:e,order:o}}catch{return{map:{},order:[]}}}function Et(r){try{let t=P.readFileSync(r,"utf-8"),n=t.match(/"id"\s*"([^"]+)"/i),e=t.match(/"name"\s*"([^"]+)"/i),o=n?n[1]:null,s=e?e[1]:null;return{id:o,name:s}}catch{return{id:null,name:null}}}function Ce(r,t){let n=['"ModList"',"{"];for(let[e,o]of Object.entries(t))n.push(`	"${e}"		"${o?"1":"0"}"`);n.push("}"),P.mkdirSync(r,{recursive:!0}),P.writeFileSync(C.join(r,"mods.vdf"),n.join(`
+`),"utf-8")}function Jt(r,t,n){let e=new Set,o=['"ModList"',"{"],s=i=>{o.push(`	"${i}"		"${n[i]?"1":"0"}"`),e.add(i)};for(let i of Array.isArray(t)?t:[])Object.prototype.hasOwnProperty.call(n,i)&&!e.has(i)&&s(i);for(let i of Object.keys(n))e.has(i)||s(i);o.push("}"),P.mkdirSync(r,{recursive:!0}),P.writeFileSync(C.join(r,"mods.vdf"),o.join(`
+`),"utf-8")}async function Ae(r,{installDir:t}){try{let n=C.join(t,"mods"),{map:e,order:o}=vt(n),s=[],i=P.existsSync(n)?P.readdirSync(n,{withFileTypes:!0}):[];for(let d of i){if(!d.isDirectory())continue;let c=C.join(n,d.name),f=C.join(c,"manifest.json"),u=C.join(c,"mod.vdf"),l=C.join(c,"icon.png"),y=null;try{y=JSON.parse(P.readFileSync(f,"utf-8"))}catch{}let m=Et(u),h=m.id||y?.name||d.name,w=null;try{w=`data:image/png;base64,${P.readFileSync(l).toString("base64")}`}catch{}s.push({id:h,name:y?.name||m.name||d.name,folder:d.name,version:y?.version_number||null,description:y?.description||"",enabled:h?!!e[h]:!1,hasManifest:P.existsSync(f),iconDataUrl:w})}let a=new Map(o.map((d,c)=>[d,c]));return s.sort((d,c)=>{let f=d.id!=null&&a.has(d.id)?a.get(d.id):Number.MAX_SAFE_INTEGER,u=c.id!=null&&a.has(c.id)?a.get(c.id):Number.MAX_SAFE_INTEGER;return f!==u?f-u:String(d.name||"").localeCompare(String(c.name||""))}),{ok:!0,mods:s}}catch(n){return{ok:!1,error:String(n?.message||n)}}}async function De(r,{installDir:t,name:n,enabled:e}){try{let o=C.join(t,"mods"),{map:s}=vt(o);return s[n]=!!e,Jt(o,void 0,s),{ok:!0}}catch(o){return{ok:!1,error:String(o?.message||o)}}}async function be(r,{installDir:t,orderIds:n}){try{let e=C.join(t,"mods"),{map:o,order:s}=vt(e),i=Array.isArray(n)?n.filter(f=>typeof f=="string"&&f&&Object.prototype.hasOwnProperty.call(o,f)):[],a=new Set(i),d=s.filter(f=>Object.prototype.hasOwnProperty.call(o,f)&&!a.has(f)),c=[...i,...d];return Jt(e,c,o),{ok:!0}}catch(e){return{ok:!1,error:String(e?.message||e)}}}async function _e(r,{installDir:t,folder:n}){try{let e=C.join(t,"mods"),o=C.join(e,n);return P.rmSync(o,{recursive:!0,force:!0}),{ok:!0}}catch(e){return{ok:!1,error:String(e?.message||e)}}}async function Le(r,{query:t}){let n="https://thunderstore.io/c/r5valkyrie/api/v1/package-listing-index/",e=s=>new Promise((i,a)=>{Ot.get(s,{headers:{"User-Agent":"R5Valkyrie-Launcher/1.0",Accept:"*/*"}},c=>{if(c.statusCode&&c.statusCode>=300&&c.statusCode<400&&c.headers.location){c.resume();let u=c.headers.location.startsWith("http")?c.headers.location:new URL(c.headers.location,s).toString();return i(e(u))}if(c.statusCode!==200)return c.resume(),a(new Error(`HTTP ${c.statusCode}`));let f=[];c.on("data",u=>f.push(Buffer.isBuffer(u)?u:Buffer.from(u))),c.on("end",()=>i(Buffer.concat(f)))}).on("error",a)}),o=s=>{try{return Pe.gunzipSync(s)}catch{return s}};try{let s=await e(n),i=JSON.parse(o(s).toString("utf8")),a=Array.isArray(i)?i:[],d=6,c=[],f=0;await Promise.all(Array.from({length:d}).map(async()=>{for(;f<a.length;){let l=f++,y=a[l];try{let m=await e(y),h=o(m).toString("utf8"),w=JSON.parse(h);Array.isArray(w)&&c.push(...w)}catch{}}}));let u=c;if(t&&String(t).trim()){let l=String(t).toLowerCase();u=u.filter(y=>String(y?.name||"").toLowerCase().includes(l)||String(y?.full_name||"").toLowerCase().includes(l))}return{ok:!0,mods:u}}catch(s){try{let a=await e("https://thunderstore.io/c/r5valkyrie/api/v1/package/"),d=JSON.parse(a.toString("utf8")),c=Array.isArray(d)?d:[];if(t&&String(t).trim()){let f=String(t).toLowerCase();c=c.filter(u=>String(u?.name||"").toLowerCase().includes(f)||String(u?.full_name||"").toLowerCase().includes(f))}return{ok:!0,mods:c}}catch(i){return{ok:!1,error:String(i?.message||s?.message||i||s)}}}}async function $e(r,{installDir:t,name:n,downloadUrl:e}){if(!e.startsWith("https://thunderstore.io"))return{ok:!1,error:"Download URL does not point to Thunderstore"};try{let o=String(n||"").trim();if(!o)return{ok:!1,error:"Invalid mod name"};if(Pt.has(o))return{ok:!0};Pt.add(o);let s=C.join(t,"mods");P.mkdirSync(s,{recursive:!0});let i=C.join(s,`.__mod_${Date.now()}.zip`),a=(u,l=0)=>new Promise((y,m)=>{if(l>5)return m(new Error("Too many redirects"));let h=P.createWriteStream(i);Ot.get(u,{headers:{"User-Agent":"R5Valkyrie-Launcher/1.0",Accept:"application/octet-stream,*/*;q=0.8",Referer:"https://thunderstore.io/",Connection:"keep-alive"}},g=>{if(g.statusCode&&g.statusCode>=300&&g.statusCode<400&&g.headers.location){g.resume(),h.close(()=>{});try{P.unlinkSync(i)}catch{}let A=g.headers.location.startsWith("http")?g.headers.location:new URL(g.headers.location,u).toString();return y(a(A,l+1))}if(g.statusCode!==200){h.close(()=>{});try{P.unlinkSync(i)}catch{}return g.resume(),m(new Error(`HTTP ${g.statusCode}`))}let S=Number(g.headers["content-length"]||0),p=0,x=Date.now(),E=A=>{try{r?.sender?.send("mods:progress",{key:n,phase:A,received:p,total:S})}catch{}};g.on("data",A=>{p+=Buffer.isBuffer(A)?A.length:Buffer.byteLength(String(A));let R=Date.now();R-x>150&&(x=R,E("downloading"))}),g.pipe(h),h.on("finish",()=>{E("downloading"),h.close(y)})}).on("error",g=>{try{P.unlinkSync(i)}catch{}m(g)})});await a(e);try{r?.sender?.send("mods:progress",{key:n,phase:"extracting"})}catch{}let d=C.join(s,n);try{P.rmSync(d,{recursive:!0,force:!0})}catch{}P.mkdirSync(d,{recursive:!0}),await new Promise((u,l)=>{let y=Ee("powershell.exe",["-NoProfile","-NonInteractive","-Command",`Expand-Archive -LiteralPath "${i}" -DestinationPath "${d}" -Force`],{stdio:"ignore"});y.on("exit",m=>m===0?u():l(new Error(`Expand-Archive failed ${m}`))),y.on("error",l)});try{P.unlinkSync(i)}catch{}let c=null;try{let u=C.join(d,"mod.vdf");if(P.existsSync(u))c=Et(u).id;else{let l=P.readdirSync(d,{withFileTypes:!0});for(let y of l)if(y.isDirectory()){let m=C.join(d,y.name,"mod.vdf");if(P.existsSync(m)&&(c=Et(m).id,c))break}}}catch{}let f=ve(s);f[c||n]=!0,Ce(s,f);try{r?.sender?.send("mods:progress",{key:n,phase:"done"})}catch{}return{ok:!0}}catch(o){return{ok:!1,error:String(o?.message||o)}}finally{try{Pt.delete(String(n||""))}catch{}}}async function Ie(r,{installDir:t},n){try{let e=C.join(t,"mods");if(!P.existsSync(e))return{ok:!0};let o=e;if(ut.has(o))return{ok:!0};let s=P.watch(e,{persistent:!0},(i,a)=>{let d=o;clearTimeout(pt.get(d));let c=setTimeout(()=>{try{n()?.webContents?.send("mods:changed",{installDir:t})}catch{}},300);pt.set(d,c)});return ut.set(o,s),{ok:!0}}catch(e){return{ok:!1,error:String(e?.message||e)}}}async function Re(r,{installDir:t}){try{let e=C.join(t,"mods"),o=ut.get(e);if(o){try{o.close()}catch{}ut.delete(e)}let s=pt.get(e);return s&&(clearTimeout(s),pt.delete(e)),{ok:!0}}catch(n){return{ok:!1,error:String(n?.message||n)}}}async function Te(r,{installDir:t,folder:n}){try{let e=C.join(t,"mods",n,"icon.png");return await P.promises.access(e),{ok:!0,dataUrl:`data:image/png;base64,${(await P.promises.readFile(e)).toString("base64")}`}}catch(e){return{ok:!1,error:String(e?.message||e)}}}function Gt(r,t){r.handle("mods:listInstalled",Ae),r.handle("mods:setEnabled",De),r.handle("mods:reorder",be),r.handle("mods:uninstall",_e),r.handle("mods:fetchAll",Le),r.handle("mods:install",$e),r.handle("mods:watch",(n,e)=>Ie(n,e,t)),r.handle("mods:unwatch",Re),r.handle("mods:iconDataUrl",Te)}import qe from"node:path";import He from"node:fs";import k from"node:fs";import v from"node:path";import Ue from"node:crypto";import{pipeline as Ne}from"node:stream";import{promisify as Ve}from"node:util";import At from"node:https";var Fe=Ve(Ne);function Ct(r){return String(r||"").replace(/\\/g,"/").replace(/^\/+/,"")}function Kt(r){return String(r||"").replace(/\\/g,"/").toLowerCase()}var rt=new Map,Xt=new At.Agent({keepAlive:!0,maxSockets:64,maxFreeSockets:32,keepAliveMsecs:6e4,scheduling:"fifo"});function X(r){return new Promise((t,n)=>{let e=Ue.createHash("sha256"),o=k.createReadStream(r);o.on("error",n),o.on("end",()=>t(e.digest("hex"))),o.on("data",s=>e.update(s))})}function Qt(r){k.mkdirSync(r,{recursive:!0})}function ze(r){return new Promise((t,n)=>{let e=At.get(r,{agent:Xt},o=>{if(o.statusCode!==200){n(new Error(`HTTP ${o.statusCode} for ${r}`)),o.resume();return}let s="";o.setEncoding("utf8"),o.on("data",i=>s+=i),o.on("end",()=>{try{t(JSON.parse(s))}catch(i){n(i)}})});e.on("error",n),e.setTimeout(3e4,()=>{e.destroy(),n(new Error("JSON fetch timeout"))})})}function B(r){return new Promise(t=>setTimeout(t,r))}function Yt(){return{cancelled:!1,requests:new Set}}function Dt(r){if(r){r.cancelled=!0;for(let t of r.requests)try{t.destroy(new Error("cancelled"))}catch{}r.requests.clear()}}var je=new Set(["ECONNRESET","ETIMEDOUT","EAI_AGAIN","ECONNABORTED","ENETRESET","EHOSTUNREACH","ENETUNREACH","EPIPE","ENOTFOUND","ECONNREFUSED","EHOSTDOWN","ENETDOWN"]);async function Q(r,t,n,e=1,o,s=0,i=0){return new Promise((a,d)=>{Qt(v.dirname(t));let c=k.createWriteStream(t,{flags:s>0?"a":"w"});if(o?.cancelled)return c.close(()=>{}),d(new Error("cancelled"));let f={agent:Xt};s>0&&(f.headers={Range:`bytes=${s}-`});let u=At.get(r,f,l=>{if(l.statusCode!==200&&!(s>0&&l.statusCode===206)){if(c.close(()=>{}),s>0&&l.statusCode===200){try{k.unlinkSync(t)}catch{}let p=Math.min(2e3*e,1e4);l.resume();try{u.destroy(new Error("restart:range-ignored"))}catch{}return B(p).then(()=>a(Q(r,t,n,e+1,o,0,i)))}if(k.unlink(t,()=>{}),l.statusCode&&[429,500,502,503,504].includes(l.statusCode)&&e<5){let p=Math.min(2e3*e,1e4);return l.resume(),B(p).then(()=>a(Q(r,t,n,e+1,o,s,i)))}d(new Error(`HTTP ${l.statusCode} for ${r}`)),l.resume();return}let y=Number(l.headers["content-length"]||0),m=i>0?i:s>0?s+y:y,h=0,w=Date.now(),g=e<=2?3e4:45e3,S=setInterval(()=>{if(!o?.cancelled&&Date.now()-w>g)try{let p=new Error("stalled");p.code="ETIMEDOUT",u.destroy(p)}catch{}},1e4);l.on("data",p=>{h+=p.length,w=Date.now(),n&&n(Math.min(s+h,m||s+h),m||0)}),l.on("aborted",async()=>{let p=new Promise(x=>c.close(()=>x()));try{clearInterval(S)}catch{}try{await p}catch{}if(e<5){let x=Math.min(2e3*e,1e4);return B(x).then(()=>a(Q(r,t,n,e+1,o,s+h,i)))}d(new Error("response aborted"))}),Fe(l,c).then(()=>{try{clearInterval(S)}catch{}a()}).catch(p=>{try{clearInterval(S)}catch{}d(p)})});if(o?.requests.add(u),u.on("socket",l=>{try{l.setKeepAlive(!0,6e4),l.setNoDelay(!0),l.setRecvBufferSize&&l.setRecvBufferSize(64*1024),l.setSendBufferSize&&l.setSendBufferSize(64*1024)}catch{}}),u.on("close",()=>o?.requests.delete(u)),u.setTimeout(9e4,()=>{try{let l=new Error("Request timeout");l.code="ETIMEDOUT",u.destroy(l)}catch{}}),o?.cancelled)try{u.destroy(new Error("cancelled"))}catch{}u.on("error",async l=>{let y=l?.code,m=e<8&&y&&je.has(String(y)),h=new Promise(w=>{try{c.close(()=>w())}catch{w()}});try{await h}catch{}if(m){let w=Math.min(1e3*Math.pow(1.5,e),15e3),g=Math.random()*1e3,S=w+g,p=s;try{p=k.statSync(t).size}catch{}return B(S).then(()=>a(Q(r,t,n,e+1,o,p,i)))}d(l)})})}async function Be(r,t,n){try{let e=k.statSync(r);return n&&Number(n)>0&&e.size!==Number(n)?!1:(await X(r)).toLowerCase()===String(t).toLowerCase()}catch{return!1}}async function Zt(r){let t=`${r.replace(/\/$/,"")}/checksums.json`;return ze(t)}async function We(r,t,n,e,o=4,s){let i=Ct(t.path).split("/").join(v.sep),a=v.join(n,i);if(Qt(v.dirname(a)),await Be(a,t.checksum,t.size))return e("progress:skip",{path:t.path}),{downloadComplete:!0,verificationComplete:!0,targetPath:a};if(s?.cancelled)throw new Error("cancelled");if(Array.isArray(t.parts)&&t.parts.length>0){let c=t.parts.length,f=new Array(c),u=0,l=new Set;async function y(){for(;;){let h=u++;if(h>=c)return;if(l.has(h))continue;l.add(h);let w=t.parts[h],g=Ct(w.path),S=`${r.replace(/\/$/,"")}/${g}`,p=`${a}.part${h}`;try{if(k.existsSync(p))if((await X(p)).toLowerCase()===String(w.checksum).toLowerCase()){e("progress:part",{path:t.path,part:h,totalParts:c,received:Number(w.size||0),total:Number(w.size||0)}),f[h]=p;continue}else try{k.unlinkSync(p)}catch{}}catch{}let x=0;for(;;){if(s?.cancelled)throw new Error("cancelled");x+=1;let E=0,A=0;try{let V=0;try{V=k.statSync(p).size}catch{}let T=Number(w.size||0);if(T>0&&V>T)try{k.unlinkSync(p),V=0}catch{}await Q(S,p,(b,W)=>{let q=Math.max(0,b-A);A=b;try{q>0&&e("progress:bytes",{delta:q})}catch{}E+=q;let J=T||W||0;e("progress:part",{path:t.path,part:h,totalParts:c,received:Math.min(b,J||b),total:J})},1,s,V,T)}catch{try{k.unlinkSync(p)}catch{}try{E>0&&e("progress:bytes",{delta:-E})}catch{}try{e("progress:part:reset",{path:t.path,part:h,totalParts:c})}catch{}let T=Math.min(2e3*x,1e4);await B(T);continue}if((await X(p)).toLowerCase()===String(w.checksum).toLowerCase())break;try{k.unlinkSync(p)}catch{}try{E>0&&e("progress:bytes",{delta:-E})}catch{}try{e("progress:part:reset",{path:t.path,part:h,totalParts:c})}catch{}let _=Math.min(2e3*x,1e4);await B(_)}f[h]=p}}let m=Array.from({length:Math.max(1,o)},()=>y());return await Promise.all(m),{downloadComplete:!0,verificationComplete:!1,targetPath:a,mergeAndVerifyAsync:async()=>{e("progress:merge:start",{path:t.path,parts:f.length});let h=k.createWriteStream(a);for(let g=0;g<f.length;g++){let S=f[g];e("progress:merge:part",{path:t.path,part:g,totalParts:f.length}),await new Promise((p,x)=>{let E=k.createReadStream(S);E.on("error",x),h.on("error",x),E.on("end",()=>{try{k.unlinkSync(S)}catch{}p()}),E.pipe(h,{end:!1})})}await new Promise((g,S)=>{h.on("error",S),h.on("finish",g),h.end()});try{e("progress:merge:done",{path:t.path})}catch{}if(e("progress:verify",{path:t.path}),(await X(a)).toLowerCase()!==String(t.checksum).toLowerCase())throw new Error(`Checksum mismatch for ${t.path}`);try{let g=k.readdirSync(v.dirname(a)),S=v.basename(a);g.forEach(p=>{if(p.startsWith(S+".part"))try{k.unlinkSync(v.join(v.dirname(a),p))}catch{}})}catch{}return{verificationComplete:!0,targetPath:a,filePath:t.path}}}}else{let c=Ct(t.path),f=`${r.replace(/\/$/,"")}/${c}`,u=0;for(;;){u+=1;let l=0,y=0,m=`${a}.download`,h=0;try{h=k.statSync(m).size}catch{}let w=Number(t.size||0);if(w>0&&h>w)try{k.unlinkSync(m),h=0}catch{}if(await Q(f,m,(S,p)=>{let x=Math.max(0,S-y);y=S;try{x>0&&e("progress:bytes",{delta:x})}catch{}l+=x;let E=w||p||0;e("progress:file",{path:t.path,received:Math.min(S,E||S),total:E})},1,s,h,w),(await X(m)).toLowerCase()===String(t.checksum).toLowerCase())break;try{k.unlinkSync(m)}catch{}try{l>0&&e("progress:bytes",{delta:-l})}catch{}if(u>=2)throw new Error(`Checksum mismatch for ${t.path}`)}try{k.unlinkSync(a)}catch{}k.renameSync(`${a}.download`,a)}if(e("progress:verify",{path:t.path}),(await X(a)).toLowerCase()!==String(t.checksum).toLowerCase())throw new Error(`Checksum mismatch for ${t.path}`);return{downloadComplete:!0,verificationComplete:!0,targetPath:a}}async function te(r,t,n,e,o=!1,s=4,i=4,a,d){let c=new Set,f=[];for(let S of t.files||[]){if(!(o||!S.optional))continue;let p=Kt(S.path);p&&(c.has(p)||(c.add(p),f.push(S)))}let u=f.filter(S=>!(Array.isArray(S.parts)&&S.parts.length>0)),l=f.filter(S=>Array.isArray(S.parts)&&S.parts.length>0),y=u.concat(l),m=y.length,h=0,w=y.reduce((S,p)=>{let x=Number(p.size||0);return x>0?S+x:Array.isArray(p.parts)&&p.parts.length?S+p.parts.reduce((E,A)=>E+Number(A.size||0),0):S},0);try{e("progress:bytes:total",{totalBytes:w})}catch{}async function g(S,p,x,E=!1){let A=0,R=[];async function _(){for(;;){for(;d&&d();)if(await B(100),a?.cancelled)return;let T=A++;if(T>=S.length)return;let b=S[T],W=p+T,q=Kt(b.path),J=async()=>We(r,b,n,e,i,a),_t=!1,at=rt.get(q);at?_t=!0:(e("progress:start",{index:W,total:m,path:b.path,completed:h}),at=(async()=>{try{let j=await J();if(E&&j?.mergeAndVerifyAsync){let $t=j.mergeAndVerifyAsync().then(L=>(L?.filePath&&(h+=1,e("progress:done",{index:W,total:m,path:L.filePath,completed:h})),L)).catch(L=>{if(String(L?.message||L||"error").includes("cancelled"))throw L;try{let U=v.join(n,b.path.replace(/\\/g,v.sep));try{k.unlinkSync(U)}catch{}try{let $=k.readdirSync(v.dirname(U)),F=v.basename(U);$.forEach(it=>{if(it.startsWith(F+".part"))try{k.unlinkSync(v.join(v.dirname(U),it))}catch{}})}catch{}return J().then($=>$.mergeAndVerifyAsync?$.mergeAndVerifyAsync().then(F=>(F?.filePath&&(h+=1,e("progress:done",{index:W,total:m,path:F.filePath,completed:h})),F)):$)}catch(U){try{e("progress:error",{path:b.path,message:String(U?.message||U)})}catch{}throw U}});return R.push($t),{...j,skipProgressDone:!0}}return j}catch(j){if(String(j?.message||j||"error").includes("cancelled"))throw j;try{let L=v.join(n,b.path.replace(/\\/g,v.sep));try{k.unlinkSync(L)}catch{}let et=await J();if(E&&et?.mergeAndVerifyAsync){let U=et.mergeAndVerifyAsync().then($=>($?.filePath&&(h+=1,e("progress:done",{index:W,total:m,path:$.filePath,completed:h})),$)).catch($=>{try{let F=v.join(n,b.path.replace(/\\/g,v.sep)),it=k.readdirSync(v.dirname(F)),pe=v.basename(F);it.forEach(It=>{if(It.startsWith(pe+".part"))try{k.unlinkSync(v.join(v.dirname(F),It))}catch{}})}catch{}try{e("progress:error",{path:b.path,message:String($?.message||$)})}catch{}throw $});return R.push(U),{...et,skipProgressDone:!0}}return et}catch(L){try{e("progress:error",{path:b.path,message:String(L?.message||L)})}catch{}throw L}}})(),rt.set(q,at));let Lt;try{Lt=await at}finally{_t||rt.delete(q)}Lt?.skipProgressDone||(h+=1,e("progress:done",{index:W,total:m,path:b.path,completed:h})),await B(10)}}let V=Array.from({length:Math.max(1,x)},()=>_());await Promise.all(V),R.length>0&&await Promise.all(R)}if(await g(u,0,Math.max(1,s)),rt.size>0)try{await Promise.all([...rt.values()])}catch{}await g(l,u.length,1,!0)}var H=null,nt=!1;function ee(r,t){r.handle("download:checksums",async(n,{baseUrl:e})=>Zt(e)),r.handle("download:all",async(n,{baseUrl:e,checksums:o,installDir:s,includeOptional:i,concurrency:a,partConcurrency:d,channelName:c,mode:f})=>{let u=(l,y)=>n.sender.send(l,y);try{if(!(String(f||"install").toLowerCase()==="install")){let y=qe.join(s,"mods","mods.vdf");if(He.existsSync(y)){let h=(Array.isArray(o?.files)?o.files:[]).filter(w=>{try{return String(w?.path||w?.name||"").replace(/\\/g,"/").toLowerCase()!=="mods/mods.vdf"}catch{return!0}});o={...o||{},files:h}}}}catch{}H=Yt(),nt=!1;try{await te(e,o,s,u,!!i,Number(a)||4,Number(d)||4,H,()=>nt);try{let l=ct("channels",{})||{};l[String(c||"default")]={installDir:s,gameVersion:o?.game_version||null,gameBaseUrl:e,lastUpdatedAt:Date.now()},G("channels",l)}catch(l){console.error("Failed to persist channel info",l)}}finally{H=null}return!0}),r.handle("download:pause",async()=>{nt=!0;try{t()?.webContents.send("progress:paused",{})}catch{}return!0}),r.handle("download:resume",async()=>{nt=!1;try{t()?.webContents.send("progress:resumed",{})}catch{}return!0}),r.handle("download:cancel",async()=>{try{Dt(H)}catch{}H=null,nt=!1;try{t()?.webContents.send("progress:cancelled",{})}catch{}return!0})}function re(){if(H)try{Dt(H)}catch{}}import Y from"node:path";import ot from"node:fs";import{spawn as ne}from"node:child_process";import{app as oe}from"electron";function Me(r){return!r||typeof r!="string"?[]:(r.match(/(?:[^\s\"]+|\"[^\"]*\")+/g)||[]).map(n=>n.replace(/^\"|\"$/g,""))}async function Oe(r,{channelName:t,installDir:n,mode:e,argsString:o,winePrefix:s}){try{if(!n)throw new Error("Missing installDir");let i=String(e).toUpperCase()==="SERVER"?"r5apex_ds.exe":"r5apex.exe",a=Y.join(n,i);await ot.promises.access(a).catch(()=>{throw new Error(`Executable not found: ${a}`)});let d=Me(o),c;if(process.platform==="win32")c=ne(a,d,{cwd:n,detached:!0,stdio:"ignore",env:{...process.env}});else if(process.platform==="linux"){let f=s||Y.join(oe.getPath("home"),"Games","R5VLibrary","wineprefix");d.unshift(a);let u=Y.join(oe.getPath("home"),".local","share","Steam"),l=Y.join(f,"drive_c","Program Files (x86)","Steam");ot.mkdirSync(l,{recursive:!0}),["GameOverlayRenderer64.dll","steamclient64.dll","steamclient.dll"].forEach(m=>{let h=Y.join(u,m),w=Y.join(l,m);if(!ot.existsSync(h))throw new Error(`Steam is not installed correctly!
+Source file not found: ${h}`);ot.existsSync(w)?console.log(`Skipped (exists): ${m}`):(ot.copyFileSync(h,w),console.log(`Copied: ${m}`))}),c=ne("umu-run",d,{cwd:n,detached:!0,stdio:"ignore",env:{...process.env,WINEPREFIX:f}})}return c.unref(),{ok:!0}}catch(i){return{ok:!1,error:String(i?.message||i)}}}function se(r){r.handle("game:launch",Oe)}import ft from"node:path";import M from"node:fs";import Je from"node:os";import{spawn as ae}from"node:child_process";async function Ge(r,{selectedChannel:t}){let n=[],e=lt().channels[t].installDir;try{if(!e)return{ok:!1,error:"No folder path provided"};let o=ft.resolve(e).replace(/\//g,"\\");if(!o||o.length<3)return{ok:!1,error:`Invalid path format: ${e}`};let s=o.length>=3&&o[1]===":"?o.slice(3):o;if(/[<>:"|?*]/.test(s))return{ok:!1,error:`Path contains invalid characters: ${o}`};let d=Je.userInfo().username,c=ft.join(process.resourcesPath,"elevate.exe"),f=M.existsSync(c);try{M.mkdirSync(o,{recursive:!0})}catch{}let u=[{exe:"icacls",args:[o,"/grant",`${d}:F`,"/t","/q"],desc:`Grant permissions to current user (${d})`},{exe:"icacls",args:[o,"/grant","Everyone:F","/t","/q"],desc:"Grant Everyone full permissions"},{exe:"icacls",args:[o,"/grant","Users:F","/t","/q"],desc:"Grant Users full permissions"}];for(let l=0;l<u.length;l++){let{exe:y,args:m,desc:h}=u[l];try{let w=await new Promise((g,S)=>{let p,x,E;f?(x=[c,["-wait",y,...m]],E={stdio:["ignore","pipe","pipe"],windowsHide:!0}):(x=[y,m],E={stdio:["ignore","pipe","pipe"],windowsHide:!0});try{p=ae(x[0],x[1],E)}catch(_){console.error("[Permission Fix] Spawn failed:",_),S(_);return}let A="",R="";p.stdout&&p.stdout.on("data",_=>{A+=_.toString()}),p.stderr&&p.stderr.on("data",_=>{R+=_.toString()}),p.on("exit",(_,V)=>{g({code:_,stdout:A,stderr:R,signal:V})}),p.on("error",_=>{S(_)}),setTimeout(()=>{try{p.kill()}catch{}S(new Error("Command timeout"))},3e4)});if(w.code!==0){let g=w.stderr||w.stdout||"Unknown error";if(!(g.includes("duplicate")||g.includes("already")||w.code===1332)){let p=`${h} failed (code ${w.code}): ${g}`;n.push(p)}}}catch(w){let g=`${h} error: ${w.message}`;n.push(g)}}try{let l=ft.join(o,"test_write.tmp");M.writeFileSync(l,"test"),M.unlinkSync(l)}catch(l){let y=`Write test failed: ${l.message}`;n.push(y);try{if(M.mkdirSync(o,{recursive:!0}),(await new Promise(h=>{let w=ae("cmd",["/c",`icacls "${o}" /grant "%USERNAME%":F /q`],{stdio:"ignore",windowsHide:!0});w.on("exit",g=>h({code:g})),w.on("error",()=>h({code:1})),setTimeout(()=>{try{w.kill()}catch{}h({code:1})},5e3)})).code===0){let h=ft.join(o,"test_write2.tmp");return M.writeFileSync(h,"test"),M.unlinkSync(h),{ok:!0,warnings:n}}}catch{}return{ok:!1,error:y,details:n}}return n.length>0?{ok:!0,warnings:n}:{ok:!0}}catch(o){let s=`Unexpected error: ${o?.message||o}`;return n.push(s),{ok:!1,error:s,details:n}}}function ie(r){r.handle("fix-folder-permissions",Ge)}import{protocol as ce,app as Ke}from"electron";import z from"node:path";import Xe from"node:fs";import{fileURLToPath as Qe}from"node:url";var Ye=Qe(import.meta.url),Ze=z.dirname(Ye);function le(){let r=z.resolve(Ke.getAppPath(),".."),t=z.join(r,"cache");ce.registerFileProtocol("r5v",(n,e)=>{try{let o=new URL(n.url),s=decodeURIComponent(o.pathname).replace(/^\//,""),i=z.normalize(z.join(t,s));e({path:i})}catch{e({error:-6})}})}function de(){let r=z.join(Ze,"..","..","dist");ce.interceptFileProtocol("file",(t,n)=>{try{let e=new URL(t.url),o=decodeURIComponent(e.pathname),s=o.replace(/\\/g,"/"),i=s.indexOf("/_astro/");if(i!==-1){let a=s.substring(i+8),d=z.join(r,"_astro",a);return n({path:d})}if(s.endsWith("/favicon.svg")||/\/favicon\.svg$/i.test(s)){let a=z.join(r,"favicon.svg");return n({path:a})}if(s.startsWith("/")){let a=s.replace(/^\/+/,""),d=z.join(r,a);try{return Xe.accessSync(d),n({path:d})}catch{}}return n({path:o})}catch{return n({error:-6})}})}import{app as Z,dialog as he}from"electron";import D from"node:path";import I from"node:fs";import bt from"node:https";import{fileURLToPath as tr}from"node:url";var er=tr(import.meta.url),rr=D.dirname(er);function ue(r){r.handle("app:getVersion",async()=>{try{return Z.getVersion()}catch{return"0.0.0"}}),r.handle("app:getBaseDir",async()=>{try{return D.resolve(Z.getAppPath(),"..")}catch{return""}}),r.handle("app:getLauncherInstallRoot",async()=>{try{let t=process.env.LOCALAPPDATA||D.join(Z.getPath("home"),"AppData","Local");return D.join(t,"Programs","r5vlauncher")}catch{return""}}),r.handle("settings:get",()=>lt()),r.handle("settings:set",(t,{key:n,value:e})=>G(n,e)),r.handle("select-directory",async()=>{let t=await he.showOpenDialog({properties:["openDirectory","createDirectory"]});return t.canceled||t.filePaths.length===0?null:t.filePaths[0]}),r.handle("select-file",async(t,{filters:n})=>{let e=await he.showOpenDialog({properties:["openFile"],filters:Array.isArray(n)?n:void 0});return e.canceled||e.filePaths.length===0?null:e.filePaths[0]}),r.handle("default-install-dir",(t,{channelName:n})=>{let e="";if(process.platform==="win32"){let o=process.env.LOCALAPPDATA||D.join(Z.getPath("home"),"AppData","Local");e=D.join(o,"Programs","R5VLibrary")}else if(process.platform==="linux"){let o=D.join(Z.getPath("home"),"Games");e=D.join(o,"R5VLibrary")}return n?D.join(e,n):e}),r.handle("launcher:config",async(t,{url:n})=>(o=>new Promise((s,i)=>{bt.get(o,a=>{if(a.statusCode!==200){i(new Error(`HTTP ${a.statusCode} for ${o}`)),a.resume();return}let d="";a.setEncoding("utf8"),a.on("data",c=>d+=c),a.on("end",()=>{try{s(JSON.parse(d))}catch(c){i(c)}})}).on("error",i)}))(n)),r.handle("eula:get",async()=>{let t="https://playvalkyrie.org/api/eula",n=e=>new Promise((o,s)=>{bt.get(e,i=>{if(i.statusCode!==200){s(new Error(`HTTP ${i.statusCode}`)),i.resume();return}let a=[];i.on("data",d=>a.push(Buffer.isBuffer(d)?d:Buffer.from(d))),i.on("end",()=>{try{let d=Buffer.concat(a).toString("utf8");o(JSON.parse(d))}catch(d){s(d)}})}).on("error",s)});try{return{ok:!0,json:await n(t)}}catch(e){return{ok:!1,error:String(e?.message||e)}}}),r.handle("video:cache",async(t,{filename:n})=>{try{let o=`https://blaze.playvalkyrie.org/video_backgrounds/${n}`,s=D.join(D.resolve(Z.getAppPath(),".."),"cache"),i=D.join(s,"videos"),a=D.join(i,n),d=D.join(rr,"..","..","dist"),c=D.join(d,n);await I.promises.mkdir(i,{recursive:!0});try{if((await I.promises.stat(a)).size>0){try{await I.promises.access(c)}catch{await I.promises.copyFile(a,c)}return`./${n}`}else await I.promises.unlink(a)}catch{}return await new Promise((f,u)=>{let l=I.createWriteStream(a),y=bt.get(o,m=>{if(m.statusCode!==200){l.close(()=>{}),I.unlink(a,()=>{}),u(new Error(`HTTP ${m.statusCode} for video ${n}`)),m.resume();return}m.pipe(l),l.on("finish",()=>l.close(f)),l.on("error",h=>{I.unlink(a,()=>u(h))})});y.on("error",m=>{l.close(()=>{}),I.unlink(a,()=>{}),u(m)}),y.setTimeout(3e4,()=>{y.destroy(),l.close(()=>{}),I.unlink(a,()=>{}),u(new Error(`Video download timeout for ${n}`))})}),await I.promises.copyFile(a,c),`./${n}`}catch(e){throw console.error(`Video cache error for ${n}:`,e),e}}),r.handle("fs:is-installed-in-dir",async(t,{path:n})=>{let e=!1,o=!1;try{await I.promises.access(D.join(n,"r5apex.exe")),e=!0}catch{}try{await I.promises.access(D.join(n,"r5apex_ds.exe")),o=!0}catch{}return e||o})}var mt;function tt(){return mt}var or=zt(tt);or||process.exit(0);st.on("window-all-closed",()=>{process.platform!=="darwin"&&st.quit()});st.whenReady().then(async()=>{jt(),le(),de(),mt=await St(),Ht(tt),Bt();try{mt?.webContents?.once("did-finish-load",()=>Ft(tt))}catch{}st.on("activate",()=>{nr.getAllWindows().length===0&&St().then(r=>{mt=r})})});Mt(O);ue(O);Gt(O,tt);ee(O,tt);se(O);ie(O);Nt(O,tt);st.on("before-quit",()=>{re()});
