+++ electron/modules/deeplinks.js
@@ -1,0 +1,1 @@
+import{app as r}from"electron";var a="r5v",i=[];function f(e){try{return(e||[]).filter(n=>typeof n=="string"&&n.startsWith(`${a}://`))}catch{return[]}}function l(e,n){try{let t=new URL(e);if(t.hostname==="mod"&&t.pathname==="/install"){let c=t.searchParams.get("name")||"",p=t.searchParams.get("version")||"",s=t.searchParams.get("downloadUrl")?.split(",")||"",o=n();if(o&&o.webContents&&!o.webContents.isLoading())try{o.webContents.send("deeplink:mod-install",{url:e,name:c,version:p,downloadUrls:s})}catch{}else i.push(e)}}catch{}}function h(e){for(;i.length;){let n=i.shift();l(n,e)}}function d(e){return r.requestSingleInstanceLock()?(r.on("second-instance",(t,c)=>{try{f(c).forEach(o=>l(o,e));let s=e();s&&(s.isMinimized()&&s.restore(),s.focus())}catch{}}),r.on("open-url",(t,c)=>{t.preventDefault(),l(c,e)}),!0):(r.quit(),!1)}function m(){try{process.defaultApp&&process.platform==="win32"?r.setAsDefaultProtocolClient(a,process.execPath,[process.argv[1]]):r.setAsDefaultProtocolClient(a)}catch{}}function k(){try{f(process.argv).forEach(e=>i.push(e))}catch{}}export{a as deeplinkScheme,f as extractDeeplinks,h as flushDeeplinks,l as handleDeeplink,k as queueInitialDeeplinks,m as registerProtocolClient,d as setupDeeplinkHandling};
