+++ electron/modules/game.js
@@ -1,0 +1,2 @@
+import t from"node:path";import o from"node:fs";import{spawn as u}from"node:child_process";import{app as f}from"electron";function y(e){return!e||typeof e!="string"?[]:(e.match(/(?:[^\s\"]+|\"[^\"]*\")+/g)||[]).map(r=>r.replace(/^\"|\"$/g,""))}async function E(e,{channelName:g,installDir:r,mode:x,argsString:S,winePrefix:p}){try{if(!r)throw new Error("Missing installDir");let n=String(x).toUpperCase()==="SERVER"?"r5apex_ds.exe":"r5apex.exe",s=t.join(r,n);await o.promises.access(s).catch(()=>{throw new Error(`Executable not found: ${s}`)});let a=y(S),i;if(process.platform==="win32")i=u(s,a,{cwd:r,detached:!0,stdio:"ignore",env:{...process.env}});else if(process.platform==="linux"){let l=p||t.join(f.getPath("home"),"Games","R5VLibrary","wineprefix");a.unshift(s);let w=t.join(f.getPath("home"),".local","share","Steam"),d=t.join(l,"drive_c","Program Files (x86)","Steam");o.mkdirSync(d,{recursive:!0}),["GameOverlayRenderer64.dll","steamclient64.dll","steamclient.dll"].forEach(c=>{let m=t.join(w,c),h=t.join(d,c);if(!o.existsSync(m))throw new Error(`Steam is not installed correctly!
+Source file not found: ${m}`);o.existsSync(h)?console.log(`Skipped (exists): ${c}`):(o.copyFileSync(m,h),console.log(`Copied: ${c}`))}),i=u("umu-run",a,{cwd:r,detached:!0,stdio:"ignore",env:{...process.env,WINEPREFIX:l}})}return i.unref(),{ok:!0}}catch(n){return{ok:!1,error:String(n?.message||n)}}}function $(e){e.handle("game:launch",E)}export{$ as setupGameIPC};
