+++ electron/modules/ipc-handlers.js
@@ -1,0 +1,1 @@
+import{app as f,dialog as _}from"electron";import s from"node:path";import i from"node:fs";import w from"node:https";import{fileURLToPath as L}from"node:url";import y from"node:fs";import $ from"node:path";import{app as x}from"electron";var D=x.getPath("userData"),S=$.join(D,"settings.json");function A(){try{let o=y.readFileSync(S,"utf-8");return JSON.parse(o)}catch{return{}}}function C(o){y.mkdirSync(D,{recursive:!0}),y.writeFileSync(S,JSON.stringify(o,null,2),"utf-8")}function v(o,n){let t=A();t[o]=n,C(t)}function k(){return A()}var T=L(import.meta.url),b=s.dirname(T);function q(o){o.handle("app:getVersion",async()=>{try{return f.getVersion()}catch{return"0.0.0"}}),o.handle("app:getBaseDir",async()=>{try{return s.resolve(f.getAppPath(),"..")}catch{return""}}),o.handle("app:getLauncherInstallRoot",async()=>{try{let n=process.env.LOCALAPPDATA||s.join(f.getPath("home"),"AppData","Local");return s.join(n,"Programs","r5vlauncher")}catch{return""}}),o.handle("settings:get",()=>k()),o.handle("settings:set",(n,{key:t,value:e})=>v(t,e)),o.handle("select-directory",async()=>{let n=await _.showOpenDialog({properties:["openDirectory","createDirectory"]});return n.canceled||n.filePaths.length===0?null:n.filePaths[0]}),o.handle("select-file",async(n,{filters:t})=>{let e=await _.showOpenDialog({properties:["openFile"],filters:Array.isArray(t)?t:void 0});return e.canceled||e.filePaths.length===0?null:e.filePaths[0]}),o.handle("default-install-dir",(n,{channelName:t})=>{let e="";if(process.platform==="win32"){let a=process.env.LOCALAPPDATA||s.join(f.getPath("home"),"AppData","Local");e=s.join(a,"Programs","R5VLibrary")}else if(process.platform==="linux"){let a=s.join(f.getPath("home"),"Games");e=s.join(a,"R5VLibrary")}return t?s.join(e,t):e}),o.handle("launcher:config",async(n,{url:t})=>(a=>new Promise((u,c)=>{w.get(a,r=>{if(r.statusCode!==200){c(new Error(`HTTP ${r.statusCode} for ${a}`)),r.resume();return}let l="";r.setEncoding("utf8"),r.on("data",p=>l+=p),r.on("end",()=>{try{u(JSON.parse(l))}catch(p){c(p)}})}).on("error",c)}))(t)),o.handle("eula:get",async()=>{let n="https://playvalkyrie.org/api/eula",t=e=>new Promise((a,u)=>{w.get(e,c=>{if(c.statusCode!==200){u(new Error(`HTTP ${c.statusCode}`)),c.resume();return}let r=[];c.on("data",l=>r.push(Buffer.isBuffer(l)?l:Buffer.from(l))),c.on("end",()=>{try{let l=Buffer.concat(r).toString("utf8");a(JSON.parse(l))}catch(l){u(l)}})}).on("error",u)});try{return{ok:!0,json:await t(n)}}catch(e){return{ok:!1,error:String(e?.message||e)}}}),o.handle("video:cache",async(n,{filename:t})=>{try{let a=`https://blaze.playvalkyrie.org/video_backgrounds/${t}`,u=s.join(s.resolve(f.getAppPath(),".."),"cache"),c=s.join(u,"videos"),r=s.join(c,t),l=s.join(b,"..","..","dist"),p=s.join(l,t);await i.promises.mkdir(c,{recursive:!0});try{if((await i.promises.stat(r)).size>0){try{await i.promises.access(p)}catch{await i.promises.copyFile(r,p)}return`./${t}`}else await i.promises.unlink(r)}catch{}return await new Promise((P,m)=>{let h=i.createWriteStream(r),g=w.get(a,d=>{if(d.statusCode!==200){h.close(()=>{}),i.unlink(r,()=>{}),m(new Error(`HTTP ${d.statusCode} for video ${t}`)),d.resume();return}d.pipe(h),h.on("finish",()=>h.close(P)),h.on("error",j=>{i.unlink(r,()=>m(j))})});g.on("error",d=>{h.close(()=>{}),i.unlink(r,()=>{}),m(d)}),g.setTimeout(3e4,()=>{g.destroy(),h.close(()=>{}),i.unlink(r,()=>{}),m(new Error(`Video download timeout for ${t}`))})}),await i.promises.copyFile(r,p),`./${t}`}catch(e){throw console.error(`Video cache error for ${t}:`,e),e}}),o.handle("fs:is-installed-in-dir",async(n,{path:t})=>{let e=!1,a=!1;try{await i.promises.access(s.join(t,"r5apex.exe")),e=!0}catch{}try{await i.promises.access(s.join(t,"r5apex_ds.exe")),a=!0}catch{}return e||a})}export{q as setupBasicIPC};
