+++ electron/modules/mods.js
@@ -1,0 +1,3 @@
+import O from"node:https";import d from"node:fs";import h from"node:path";import F from"node:zlib";import{spawn as E}from"node:child_process";var j=new Set,b=new Map,v=new Map;function M(i){try{let a=h.join(i,"mods.vdf"),e=d.readFileSync(a,"utf-8"),t={},n=/"([^"]+)"\s*"([01])"/g,r;for(;r=n.exec(e);){let o=r[1];o&&o!=="ModList"&&(t[o]=r[2]==="1")}return t}catch{return{}}}function A(i){try{let a=h.join(i,"mods.vdf"),e=d.readFileSync(a,"utf-8"),t={},n=[],r=/"([^"]+)"\s*"([01])"/g,o;for(;o=r.exec(e);){let u=o[1];!u||u==="ModList"||(u in t||n.push(u),t[u]=o[2]==="1")}return{map:t,order:n}}catch{return{map:{},order:[]}}}function x(i){try{let a=d.readFileSync(i,"utf-8"),e=a.match(/"id"\s*"([^"]+)"/i),t=a.match(/"name"\s*"([^"]+)"/i),n=e?e[1]:null,r=t?t[1]:null;return{id:n,name:r}}catch{return{id:null,name:null}}}function U(i,a){let e=['"ModList"',"{"];for(let[t,n]of Object.entries(a))e.push(`	"${t}"		"${n?"1":"0"}"`);e.push("}"),d.mkdirSync(i,{recursive:!0}),d.writeFileSync(h.join(i,"mods.vdf"),e.join(`
+`),"utf-8")}function T(i,a,e){let t=new Set,n=['"ModList"',"{"],r=o=>{n.push(`	"${o}"		"${e[o]?"1":"0"}"`),t.add(o)};for(let o of Array.isArray(a)?a:[])Object.prototype.hasOwnProperty.call(e,o)&&!t.has(o)&&r(o);for(let o of Object.keys(e))t.has(o)||r(o);n.push("}"),d.mkdirSync(i,{recursive:!0}),d.writeFileSync(h.join(i,"mods.vdf"),n.join(`
+`),"utf-8")}async function $(i,{installDir:a}){try{let e=h.join(a,"mods"),{map:t,order:n}=A(e),r=[],o=d.existsSync(e)?d.readdirSync(e,{withFileTypes:!0}):[];for(let c of o){if(!c.isDirectory())continue;let s=h.join(e,c.name),f=h.join(s,"manifest.json"),l=h.join(s,"mod.vdf"),p=h.join(s,"icon.png"),m=null;try{m=JSON.parse(d.readFileSync(f,"utf-8"))}catch{}let g=x(l),S=g.id||m?.name||c.name,k=null;try{k=`data:image/png;base64,${d.readFileSync(p).toString("base64")}`}catch{}r.push({id:S,name:m?.name||g.name||c.name,folder:c.name,version:m?.version_number||null,description:m?.description||"",enabled:S?!!t[S]:!1,hasManifest:d.existsSync(f),iconDataUrl:k})}let u=new Map(n.map((c,s)=>[c,s]));return r.sort((c,s)=>{let f=c.id!=null&&u.has(c.id)?u.get(c.id):Number.MAX_SAFE_INTEGER,l=s.id!=null&&u.has(s.id)?u.get(s.id):Number.MAX_SAFE_INTEGER;return f!==l?f-l:String(c.name||"").localeCompare(String(s.name||""))}),{ok:!0,mods:r}}catch(e){return{ok:!1,error:String(e?.message||e)}}}async function I(i,{installDir:a,name:e,enabled:t}){try{let n=h.join(a,"mods"),{map:r}=A(n);return r[e]=!!t,T(n,void 0,r),{ok:!0}}catch(n){return{ok:!1,error:String(n?.message||n)}}}async function N(i,{installDir:a,orderIds:e}){try{let t=h.join(a,"mods"),{map:n,order:r}=A(t),o=Array.isArray(e)?e.filter(f=>typeof f=="string"&&f&&Object.prototype.hasOwnProperty.call(n,f)):[],u=new Set(o),c=r.filter(f=>Object.prototype.hasOwnProperty.call(n,f)&&!u.has(f)),s=[...o,...c];return T(t,s,n),{ok:!0}}catch(t){return{ok:!1,error:String(t?.message||t)}}}async function B(i,{installDir:a,folder:e}){try{let t=h.join(a,"mods"),n=h.join(t,e);return d.rmSync(n,{recursive:!0,force:!0}),{ok:!0}}catch(t){return{ok:!1,error:String(t?.message||t)}}}async function R(i,{query:a}){let e="https://thunderstore.io/c/r5valkyrie/api/v1/package-listing-index/",t=r=>new Promise((o,u)=>{O.get(r,{headers:{"User-Agent":"R5Valkyrie-Launcher/1.0",Accept:"*/*"}},s=>{if(s.statusCode&&s.statusCode>=300&&s.statusCode<400&&s.headers.location){s.resume();let l=s.headers.location.startsWith("http")?s.headers.location:new URL(s.headers.location,r).toString();return o(t(l))}if(s.statusCode!==200)return s.resume(),u(new Error(`HTTP ${s.statusCode}`));let f=[];s.on("data",l=>f.push(Buffer.isBuffer(l)?l:Buffer.from(l))),s.on("end",()=>o(Buffer.concat(f)))}).on("error",u)}),n=r=>{try{return F.gunzipSync(r)}catch{return r}};try{let r=await t(e),o=JSON.parse(n(r).toString("utf8")),u=Array.isArray(o)?o:[],c=6,s=[],f=0;await Promise.all(Array.from({length:c}).map(async()=>{for(;f<u.length;){let p=f++,m=u[p];try{let g=await t(m),S=n(g).toString("utf8"),k=JSON.parse(S);Array.isArray(k)&&s.push(...k)}catch{}}}));let l=s;if(a&&String(a).trim()){let p=String(a).toLowerCase();l=l.filter(m=>String(m?.name||"").toLowerCase().includes(p)||String(m?.full_name||"").toLowerCase().includes(p))}return{ok:!0,mods:l}}catch(r){try{let u=await t("https://thunderstore.io/c/r5valkyrie/api/v1/package/"),c=JSON.parse(u.toString("utf8")),s=Array.isArray(c)?c:[];if(a&&String(a).trim()){let f=String(a).toLowerCase();s=s.filter(l=>String(l?.name||"").toLowerCase().includes(f)||String(l?.full_name||"").toLowerCase().includes(f))}return{ok:!0,mods:s}}catch(o){return{ok:!1,error:String(o?.message||r?.message||o||r)}}}}async function W(i,{installDir:a,name:e,downloadUrl:t}){if(!t.startsWith("https://thunderstore.io"))return{ok:!1,error:"Download URL does not point to Thunderstore"};try{let n=String(e||"").trim();if(!n)return{ok:!1,error:"Invalid mod name"};if(j.has(n))return{ok:!0};j.add(n);let r=h.join(a,"mods");d.mkdirSync(r,{recursive:!0});let o=h.join(r,`.__mod_${Date.now()}.zip`),u=(l,p=0)=>new Promise((m,g)=>{if(p>5)return g(new Error("Too many redirects"));let S=d.createWriteStream(o);O.get(l,{headers:{"User-Agent":"R5Valkyrie-Launcher/1.0",Accept:"application/octet-stream,*/*;q=0.8",Referer:"https://thunderstore.io/",Connection:"keep-alive"}},y=>{if(y.statusCode&&y.statusCode>=300&&y.statusCode<400&&y.headers.location){y.resume(),S.close(()=>{});try{d.unlinkSync(o)}catch{}let w=y.headers.location.startsWith("http")?y.headers.location:new URL(y.headers.location,l).toString();return m(u(w,p+1))}if(y.statusCode!==200){S.close(()=>{});try{d.unlinkSync(o)}catch{}return y.resume(),g(new Error(`HTTP ${y.statusCode}`))}let D=Number(y.headers["content-length"]||0),C=0,L=Date.now(),P=w=>{try{i?.sender?.send("mods:progress",{key:e,phase:w,received:C,total:D})}catch{}};y.on("data",w=>{C+=Buffer.isBuffer(w)?w.length:Buffer.byteLength(String(w));let _=Date.now();_-L>150&&(L=_,P("downloading"))}),y.pipe(S),S.on("finish",()=>{P("downloading"),S.close(m)})}).on("error",y=>{try{d.unlinkSync(o)}catch{}g(y)})});await u(t);try{i?.sender?.send("mods:progress",{key:e,phase:"extracting"})}catch{}let c=h.join(r,e);try{d.rmSync(c,{recursive:!0,force:!0})}catch{}d.mkdirSync(c,{recursive:!0}),await new Promise((l,p)=>{let m=E("powershell.exe",["-NoProfile","-NonInteractive","-Command",`Expand-Archive -LiteralPath "${o}" -DestinationPath "${c}" -Force`],{stdio:"ignore"});m.on("exit",g=>g===0?l():p(new Error(`Expand-Archive failed ${g}`))),m.on("error",p)});try{d.unlinkSync(o)}catch{}let s=null;try{let l=h.join(c,"mod.vdf");if(d.existsSync(l))s=x(l).id;else{let p=d.readdirSync(c,{withFileTypes:!0});for(let m of p)if(m.isDirectory()){let g=h.join(c,m.name,"mod.vdf");if(d.existsSync(g)&&(s=x(g).id,s))break}}}catch{}let f=M(r);f[s||e]=!0,U(r,f);try{i?.sender?.send("mods:progress",{key:e,phase:"done"})}catch{}return{ok:!0}}catch(n){return{ok:!1,error:String(n?.message||n)}}finally{try{j.delete(String(e||""))}catch{}}}async function V(i,{installDir:a},e){try{let t=h.join(a,"mods");if(!d.existsSync(t))return{ok:!0};let n=t;if(b.has(n))return{ok:!0};let r=d.watch(t,{persistent:!0},(o,u)=>{let c=n;clearTimeout(v.get(c));let s=setTimeout(()=>{try{e()?.webContents?.send("mods:changed",{installDir:a})}catch{}},300);v.set(c,s)});return b.set(n,r),{ok:!0}}catch(t){return{ok:!1,error:String(t?.message||t)}}}async function z(i,{installDir:a}){try{let t=h.join(a,"mods"),n=b.get(t);if(n){try{n.close()}catch{}b.delete(t)}let r=v.get(t);return r&&(clearTimeout(r),v.delete(t)),{ok:!0}}catch(e){return{ok:!1,error:String(e?.message||e)}}}async function J(i,{installDir:a,folder:e}){try{let t=h.join(a,"mods",e,"icon.png");return await d.promises.access(t),{ok:!0,dataUrl:`data:image/png;base64,${(await d.promises.readFile(t)).toString("base64")}`}}catch(t){return{ok:!1,error:String(t?.message||t)}}}function Z(i,a){i.handle("mods:listInstalled",$),i.handle("mods:setEnabled",I),i.handle("mods:reorder",N),i.handle("mods:uninstall",B),i.handle("mods:fetchAll",R),i.handle("mods:install",W),i.handle("mods:watch",(e,t)=>V(e,t,a)),i.handle("mods:unwatch",z),i.handle("mods:iconDataUrl",J)}export{Z as setupModsIPC};
