+++ electron/modules/permissions.js
@@ -1,0 +1,1 @@
+import g from"node:path";import a from"node:fs";import N from"node:os";import{spawn as P}from"node:child_process";import C from"node:fs";import I from"node:path";import{app as U}from"electron";var q=U.getPath("userData"),A=I.join(q,"settings.json");function D(){try{let u=C.readFileSync(A,"utf-8");return JSON.parse(u)}catch{return{}}}function F(){return D()}async function O(u,{selectedChannel:E}){let o=[],h=F().channels[E].installDir;try{if(!h)return{ok:!1,error:"No folder path provided"};let r=g.resolve(h).replace(/\//g,"\\");if(!r||r.length<3)return{ok:!1,error:`Invalid path format: ${h}`};let d=r.length>=3&&r[1]===":"?r.slice(3):r;if(/[<>:"|?*]/.test(d))return{ok:!1,error:`Path contains invalid characters: ${r}`};let y=N.userInfo().username,S=g.join(process.resourcesPath,"elevate.exe"),$=a.existsSync(S);try{a.mkdirSync(r,{recursive:!0})}catch{}let k=[{exe:"icacls",args:[r,"/grant",`${y}:F`,"/t","/q"],desc:`Grant permissions to current user (${y})`},{exe:"icacls",args:[r,"/grant","Everyone:F","/t","/q"],desc:"Grant Everyone full permissions"},{exe:"icacls",args:[r,"/grant","Users:F","/t","/q"],desc:"Grant Users full permissions"}];for(let n=0;n<k.length;n++){let{exe:l,args:f,desc:i}=k[n];try{let e=await new Promise((c,p)=>{let t,m,w;$?(m=[S,["-wait",l,...f]],w={stdio:["ignore","pipe","pipe"],windowsHide:!0}):(m=[l,f],w={stdio:["ignore","pipe","pipe"],windowsHide:!0});try{t=P(m[0],m[1],w)}catch(s){console.error("[Permission Fix] Spawn failed:",s),p(s);return}let x="",v="";t.stdout&&t.stdout.on("data",s=>{x+=s.toString()}),t.stderr&&t.stderr.on("data",s=>{v+=s.toString()}),t.on("exit",(s,M)=>{c({code:s,stdout:x,stderr:v,signal:M})}),t.on("error",s=>{p(s)}),setTimeout(()=>{try{t.kill()}catch{}p(new Error("Command timeout"))},3e4)});if(e.code!==0){let c=e.stderr||e.stdout||"Unknown error";if(!(c.includes("duplicate")||c.includes("already")||e.code===1332)){let t=`${i} failed (code ${e.code}): ${c}`;o.push(t)}}}catch(e){let c=`${i} error: ${e.message}`;o.push(c)}}try{let n=g.join(r,"test_write.tmp");a.writeFileSync(n,"test"),a.unlinkSync(n)}catch(n){let l=`Write test failed: ${n.message}`;o.push(l);try{if(a.mkdirSync(r,{recursive:!0}),(await new Promise(i=>{let e=P("cmd",["/c",`icacls "${r}" /grant "%USERNAME%":F /q`],{stdio:"ignore",windowsHide:!0});e.on("exit",c=>i({code:c})),e.on("error",()=>i({code:1})),setTimeout(()=>{try{e.kill()}catch{}i({code:1})},5e3)})).code===0){let i=g.join(r,"test_write2.tmp");return a.writeFileSync(i,"test"),a.unlinkSync(i),{ok:!0,warnings:o}}}catch{}return{ok:!1,error:l,details:o}}return o.length>0?{ok:!0,warnings:o}:{ok:!0}}catch(r){let d=`Unexpected error: ${r?.message||r}`;return o.push(d),{ok:!1,error:d,details:o}}}function K(u){u.handle("fix-folder-permissions",O)}export{K as setupPermissionsIPC};
