+++ electron/modules/protocols.js
@@ -1,0 +1,1 @@
+import{protocol as h,app as m}from"electron";import t from"node:path";import f from"node:fs";import{fileURLToPath as l}from"node:url";var d=l(import.meta.url),u=t.dirname(d);function j(){let r=t.resolve(m.getAppPath(),".."),p=t.join(r,"cache");h.registerFileProtocol("r5v",(e,n)=>{try{let i=new URL(e.url),o=decodeURIComponent(i.pathname).replace(/^\//,""),c=t.normalize(t.join(p,o));n({path:c})}catch{n({error:-6})}})}function R(){let r=t.join(u,"..","..","dist");h.interceptFileProtocol("file",(p,e)=>{try{let n=new URL(p.url),i=decodeURIComponent(n.pathname),o=i.replace(/\\/g,"/"),c=o.indexOf("/_astro/");if(c!==-1){let s=o.substring(c+8),a=t.join(r,"_astro",s);return e({path:a})}if(o.endsWith("/favicon.svg")||/\/favicon\.svg$/i.test(o)){let s=t.join(r,"favicon.svg");return e({path:s})}if(o.startsWith("/")){let s=o.replace(/^\/+/,""),a=t.join(r,s);try{return f.accessSync(a),e({path:a})}catch{}}return e({path:i})}catch{return e({error:-6})}})}export{R as interceptFileProtocol,j as registerCacheProtocol};
