+++ electron/modules/window.js
@@ -1,0 +1,2 @@
+var g=(t=>typeof require<"u"?require:typeof Proxy<"u"?new Proxy(t,{get:(e,n)=>(typeof require<"u"?require:e)[n]}):t)(function(t){if(typeof require<"u")return require.apply(this,arguments);throw Error('Dynamic require of "'+t+'" is not supported')});import{BrowserWindow as y,shell as x}from"electron";import s from"node:path";import{fileURLToPath as V,pathToFileURL as z}from"node:url";import l from"node:fs";import _ from"node:path";import{app as v}from"electron";var f=v.getPath("userData"),h=_.join(f,"settings.json");function w(){try{let t=l.readFileSync(h,"utf-8");return JSON.parse(t)}catch{return{}}}function R(t){l.mkdirSync(f,{recursive:!0}),l.writeFileSync(h,JSON.stringify(t,null,2),"utf-8")}function u(t,e=void 0){return w()[t]??e}function p(t,e){let n=w();n[t]=e,R(n)}var D=V(import.meta.url),c=s.dirname(D);function B(){let t=process.env.VITE_DEV_SERVER_URL;if(t)return t;let e=s.join(c,"..","dist","index.html");return z(e).toString()}async function N(){let t=u("windowState",{width:1150,height:800,x:void 0,y:void 0,isMaximized:!1}),e=new y({width:t.width,height:t.height,x:t.x,y:t.y,minWidth:1150,minHeight:800,frame:!1,titleBarStyle:"hidden",webPreferences:{preload:s.join(c,"preload.cjs"),contextIsolation:!0,nodeIntegration:!1},show:!1,icon:s.join(c,"..","public","icon.png")}),n=process.env.VITE_DEV_SERVER_URL;if(n)await e.loadURL(n);else{let i=s.join(c,"..","dist","index.html");await e.loadFile(i)}e.webContents.setWindowOpenHandler(({url:i})=>{try{x.openExternal(i)}catch{}return{action:"deny"}}),e.webContents.on("will-navigate",(i,o)=>{let a=o.startsWith("file:"),d=!!process.env.VITE_DEV_SERVER_URL&&o.startsWith(String(process.env.VITE_DEV_SERVER_URL));if(!a&&!d){i.preventDefault();try{x.openExternal(o)}catch{}}});let m=()=>{e&&(e.isVisible()||e.show(),e.isFocused()||e.focus())};t.isMaximized&&e.maximize(),e.once("ready-to-show",m),e.webContents.once("did-finish-load",m);let S=setTimeout(m,1e3);e.on("show",()=>clearTimeout(S));function r(){if(!e)return;let i=e.getBounds(),o=e.isMaximized();p("windowState",{width:i.width,height:i.height,x:i.x,y:i.y,isMaximized:o})}return e.on("resize",r),e.on("move",r),e.on("maximize",r),e.on("unmaximize",r),(!!process.env.VITE_DEV_SERVER_URL||process.env.NODE_ENV==="development")&&e.webContents.openDevTools({mode:"detach"}),e.webContents.on("did-fail-load",(i,o,a,d)=>{console.error("Load failed",{errorCode:o,errorDesc:a,validatedURL:d});let{dialog:E}=g("electron");E.showErrorBox("Load failed",`${a} (code ${o})
+URL: ${d}`),e.show()}),e}function O(t,e){t.on("window:minimize",()=>{e()?.minimize()}),t.on("window:maximize",()=>{let n=e();n&&(n.isMaximized()?n.unmaximize():n.maximize())}),t.on("window:close",()=>{e()?.close()})}export{N as createWindow,B as getAppUrl,O as setupWindowIPC};
