+++ package.json
@@ -29,7 +29,7 @@
     "@tailwindcss/vite": "^4.1.13",
     "@types/react": "^19.1.12",
     "@types/react-dom": "^19.1.9",
-    "astro": "^5.13.5",
+    "astro": "^5.14.1",
     "concurrently": "^9.2.1",
     "cross-env": "^10.0.0",
     "daisyui": "^5.1.7",
