+++ scripts/bundleElectron.cjs
@@ -11,6 +11,24 @@ function ensureDir(dir) { fs.mkdirSync(dir, { recursive: true }); }
   const outDir = path.join(projectRoot, 'electron');
   ensureDir(outDir);
   ensureDir(path.join(outDir, 'services'));
+  ensureDir(path.join(outDir, 'modules'));
+
+  // Common external modules for Node.js
+  const nodeExternals = [
+    'electron',
+    'fs', 'node:fs',
+    'path', 'node:path',
+    'https', 'node:https',
+    'http', 'node:http',
+    'child_process', 'node:child_process',
+    'os', 'node:os',
+    'zlib', 'node:zlib',
+    'crypto', 'node:crypto',
+    'stream', 'node:stream',
+    'util', 'node:util',
+    'url', 'node:url',
+    'module', 'node:module',
+  ];
 
   // Bundle main (ESM)
   await build({
@@ -22,13 +40,13 @@ function ensureDir(dir) { fs.mkdirSync(dir, { recursive: true }); }
     minify: true,
     sourcemap: false,
     logLevel: 'silent',
-    external: ['electron', 'fs', 'path', 'node:fs', 'node:path', 'https', 'node:https', 'child_process', 'node:child_process'],
+    external: nodeExternals,
   });
   // Avoid obfuscating main to keep Node/Electron semantics intact
 
   // Bundle preload (CJS maintained if needed)
   const preloadSrc = path.join(projectRoot, 'js', 'preload.cjs');
-  if (fs.existsSync(projectRoot, 'electron')) {
+  if (fs.existsSync(preloadSrc)) {
     await build({
       entryPoints: [preloadSrc],
       outfile: path.join(outDir, 'preload.cjs'),
@@ -38,7 +56,7 @@ function ensureDir(dir) { fs.mkdirSync(dir, { recursive: true }); }
       minify: true,
       sourcemap: false,
       logLevel: 'silent',
-      external: ['electron', 'fs', 'path', 'node:fs', 'node:path'],
+      external: nodeExternals,
     });
     // Avoid obfuscating preload for stability
   }
@@ -58,8 +76,38 @@ function ensureDir(dir) { fs.mkdirSync(dir, { recursive: true }); }
       minify: true,
       sourcemap: false,
       logLevel: 'silent',
-      external: ['electron', 'fs', 'path', 'node:fs', 'node:path', 'https', 'node:https', 'child_process', 'node:child_process'],
+      external: nodeExternals,
     });
-    // Avoid obfuscating services to preserve ESM exports
   }
+
+  // Bundle modules (ESM)
+  const modules = [
+    'window.js',
+    'deeplinks.js', 
+    'updater.js',
+    'mods.js',
+    'downloads.js',
+    'game.js',
+    'permissions.js',
+    'protocols.js',
+    'ipc-handlers.js'
+  ];
+  for (const m of modules) {
+    const src = path.join(projectRoot, 'js', 'modules', m);
+    if (!fs.existsSync(src)) continue;
+    const out = path.join(outDir, 'modules', m);
+    await build({
+      entryPoints: [src],
+      outfile: out,
+      bundle: true,
+      platform: 'node',
+      format: 'esm',
+      minify: true,
+      sourcemap: false,
+      logLevel: 'silent',
+      external: nodeExternals,
+    });
+  }
+
+  console.log('✅ Bundled main.js, preload.cjs, services, and modules successfully');
 })();
